### 2026-06-25 #10805 @editorialbot
[REVIEW]: PyMC-Marketing: Bayesian Marketing Mix Models and Customer Analytics in Python
**Submitting author:** @williambdean (<a href="http://orcid.org/0009-0003-6510-3545">William Dean</a>)
**Repository:** https://github.com/pymc-labs/pymc-marketing
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** 1.1.0
