### 2026-06-26 2f6de62
fix: address review feedback from cetagostini and juanitorduz
- idata_to_zarr: preserve root attrs before filter(), not self-copy after
  (DataTree.filter() does not carry over root attrs)
- mmm.py: posterior = idata["/posterior"].dataset
  (was idata.posterior, returns DataTree child, not Dataset)
- pyproject.toml: pymc-extras upper bound <0.12.0 -> <0.13.0
  (git pin resolves to 0.12.2.dev1, outside old range)
- dev notebooks: remove broken cells referencing deleted
  ContContract, ContNonContract, and deprecated fit_method API
