### 2026-06-26 866e95d
fix: revert dims.py — restore XTensorVariable guard needed in CI
- PyTensor 3.x in CI still blocks implicit XTensorVariable conversion.
  The isinstance guard with .values is required for incrementality
  vectorization to work with time-varying media models.
- Reverts commit 90afc051 for dims.py only.
