+++ pymc_marketing/mmm/dims.py
@@ -16,10 +16,11 @@
 from pytensor.graph.replace import _vectorize_node
 from pytensor.tensor import TensorLike
 from pytensor.tensor.shape import Shape_i, shape_i
+from pytensor.xtensor.type import XTensorVariable
 from xarray import DataArray
 
 # TODO: This will eventually exist in PyTensor or PyMC, remove then
-type XTensorLike = TensorLike | DataArray
+type XTensorLike = TensorLike | XTensorVariable | DataArray
 
 
 @_vectorize_node.register(Shape_i)
@@ -29,6 +30,8 @@ def _vectorize_shape_i(op: Shape_i, node, batched_x):
     [old_x] = node.inputs
     core_ndims = old_x.type.ndim
     batch_ndims = batched_x.type.ndim - core_ndims
+    if isinstance(batched_x, XTensorVariable):
+        batched_x = batched_x.values
     batched_x_shape_i = shape_i(batched_x, op.i + batch_ndims)
     if not batch_ndims:
         return [batched_x_shape_i]
