### 2026-06-26 1e2b2ed
chore: migrate to pymc>=6.0.0, arviz>=1.1.0, pytensor>=3.0.0 (#2616)
Migrate from az.InferenceData to xr.DataTree across the entire codebase.

Key changes:
- All az.InferenceData → xr.DataTree type annotations (~44 source files)
- idata.extend() → idata.update(), add_groups() → bracket assignment
- az.plot_* → azp.plot_* (arviz_plots), pm.plot_trace → azp.plot_trace
- DataTree child variable access: .posterior.VAR → .posterior["VAR"] (25 fixes)
- az.hdi(): hdi_prob= → prob=, ci_bounds → ci_bound, "higher" → "upper"
- shifted_beta_geo: restore pm.sample_posterior_predictive() with sample_vars
-
[…]
