+++ tests/clv/models/test_beta_geo.py
@@ -48,8 +48,8 @@ class TestBetaGeoModel:
         cls.T = test_data["T"]
 
         # Instantiate model with CDNOW data for testing
-        cls.model = BetaGeoModel(cls.data)
-        cls.model.build_model()
+        cls.model = BetaGeoModel()
+        cls.model.build_model(data=cls.data)
 
         # Mock an idata object for tests requiring a fitted model
         cls.N = len(cls.customer_id)
@@ -261,9 +261,9 @@ class TestBetaGeoModel:
 
         if fit_type == "map":
             map_idata = self.model.idata.copy()
-            map_idata.posterior = map_idata.posterior.isel(
-                chain=slice(None, 1), draw=slice(None, 1)
-            )
+            ds = map_idata["/posterior"].to_dataset()
+            reduced = ds.isel(chain=slice(None, 1), draw=slice(None, 1))
+            map_idata["/posterior"] = reduced
             model = self.model.build_from_idata(map_idata)
             # We expect 1000 draws to be sampled with MAP
             expected_shape = (1, 1000)
@@ -372,7 +372,7 @@ class TestBetaGeoModel:
             draws=10,
             compute_convergence_checks=False,
         )
-        assert isinstance(idata, az.InferenceData)
+        assert isinstance(idata, xr.DataTree)
         assert len(idata.posterior.chain) == 2
         assert len(idata.posterior.draw) == 10
         assert model.idata is idata
@@ -397,10 +397,12 @@ class TestBetaGeoModel:
         bg_model.build_model(data=data)
         bg_model.idata = az.from_dict(
             {
-                "a": np.full((2, 5), self.a_true),
-                "b": np.full((2, 5), self.b_true),
-                "alpha": np.full((2, 5), self.alpha_true),
-                "r": np.full((2, 5), self.r_true),
+                "posterior": {
+                    "a": np.full((2, 5), self.a_true),
+                    "b": np.full((2, 5), self.b_true),
+                    "alpha": np.full((2, 5), self.alpha_true),
+                    "r": np.full((2, 5), self.r_true),
+                }
             }
         )
 
@@ -429,10 +431,12 @@ class TestBetaGeoModel:
         bg_model.build_model(data=data)
         bg_model.idata = az.from_dict(
             {
-                "a": np.full((2, 5), self.a_true),
-                "b": np.full((2, 5), self.b_true),
-                "alpha": np.full((2, 5), self.alpha_true),
-                "r": np.full((2, 5), self.r_true),
+                "posterior": {
+                    "a": np.full((2, 5), self.a_true),
+                    "b": np.full((2, 5), self.b_true),
+                    "alpha": np.full((2, 5), self.alpha_true),
+                    "r": np.full((2, 5), self.r_true),
+                }
             }
         )
 
@@ -460,10 +464,12 @@ class TestBetaGeoModel:
         bg_model.build_model(data=data)
         bg_model.idata = az.from_dict(
             {
-                "a": np.full((2, 5), self.a_true),
-                "b": np.full((2, 5), self.b_true),
-                "alpha": np.full((2, 5), self.alpha_true),
-                "r": np.full((2, 5), self.r_true),
+                "posterior": {
+                    "a": np.full((2, 5), self.a_true),
+                    "b": np.full((2, 5), self.b_true),
+                    "alpha": np.full((2, 5), self.alpha_true),
+                    "r": np.full((2, 5), self.r_true),
+                }
             }
         )
 
@@ -491,10 +497,12 @@ class TestBetaGeoModel:
         bg_model.build_model(data=data)
         bg_model.idata = az.from_dict(
             {
-                "a": np.full((2, 5), self.a_true),
[…]
