### 2026-06-26 f9c28f2
chore: remove dead catch_warnings blocks for add_groups
- The 'fit_data not defined in InferenceData scheme' warning was
  emitted by arviz add_groups, which has been replaced by
  idata["/fit_data"] = fit_data. DataTree assignment never raises it.
- Removed dead with/catch_warnings/filterwarnings blocks in
  mixed_logit, mnl_logit, nested_logit, maxdiff.
