+++ pymc_marketing/customer_choice/maxdiff.py
@@ -1194,13 +1194,7 @@ class MaxDiffMixedLogit(ModelBuilder):
             self.idata = self.idata.drop_nodes("fit_data")
 
         fit_data = self._create_fit_data()
-        with warnings.catch_warnings():
-            warnings.filterwarnings(
-                "ignore",
-                category=UserWarning,
-                message="The group fit_data is not defined in the InferenceData scheme",
-            )
-            self.idata["/fit_data"] = fit_data
+        self.idata["/fit_data"] = fit_data
 
         self.set_idata_attrs(self.idata)
         return self.idata
