### 2026-06-26 a9c5787
chore: add pre-commit hook for notebook watermark validation
Add scripts/validate_notebook_watermark.py — a pre-commit hook that
validates every Jupyter notebook has a %watermark cell as its last cell
with required flags (-n -u -v -iv -w) and -p package list.

Add validate-watermark hook to .pre-commit-config.yaml, excluding dev/
directories.
