### 2026-06-26 b690f5e
fix: add watermark -p flags and fix missing/incomplete watermark cells
Add -p package flags to 11 notebooks missing them.
Remove empty trailing code cells from 5 notebooks.
Add watermark cells to 2 notebooks that had none.

This ensures all notebooks pass the validate-watermark pre-commit hook.
