+++ .pre-commit-config.yaml
@@ -65,6 +65,7 @@ repos:
       - id: gallery-in-sync
         name: gallery.md is in sync with gallery.yaml
         entry: python3 scripts/generate_gallery.py --check --no-thumbnails
-        language: system
+        language: python
+        additional_dependencies: [pyyaml, matplotlib]
         pass_filenames: false
         files: ^(docs/source/gallery/gallery\.(md|yaml)|docs/source/notebooks/.+\.ipynb|scripts/generate_gallery\.py)$
