### 2026-06-26 2d08f90
feat: require pymc_marketing in -p if imported by notebook
Add import-based check: if any code cell imports pymc_marketing
(import/from), the -p flag must include pymc_marketing.
Fix 14 notebooks that imported pymc_marketing but omitted it.
