### 2026-06-26 9f16d26
feat(plot): add n_samples=0 support and list hdi_prob in plot_hdi (#2650)
Original commits by pkaf in PR #2257, adapted to v1.0.0:

- 41714d07 nsamples 0 and hdi loop added
- 45f1d851 nsamples 0 and hdi loop comments addressed
- b7970ae5 nsamples 0 and hdi loop comments addressed
- e016ae66 nsamples 0 and hdi loop added
- 8ce9052b nsamples 0 and hdi loop comments addressed
- d210c4e4 nsamples comments addressed
- 27bfec22 nsamples comments addressed
- f58c9ddb plot_hdi arg corrected to hdi_prob
- 0dcb48f2 hdi_prob parameter fix
- 2b97fe38 subplot_kwargs fixed
- 9bc530a6 type hints error fixed

[…]
