### 2026-06-26 bf14944
Apply review feedback: scan scope alignment, utf-8, regex whitespace, deterministic ordering
