### 2026-06-26 b2d6928
Apply ruff formatting
