### 2026-06-26 ca731a3
refactor(budget-optimizer): BudgetOptimizer accepts pm.Model + DataTree directly
Replaces the wrapper class requirement with direct pm.Model + DataTree
inputs. Adds configurable variable/dimension names, standalone merge
utilities, and a model_validator for backward compatibility.

- BudgetOptimizer: model + idata fields replace mmm_model wrapper
- New fields: adstock_periods, channel_scales, channel_data_var,
  channel_contribution_var, date_dim, mu_effects (Sequence)
- model_validator('before') unpacks legacy wrappers transparently
- New: merge_inference_data() and merge_models_and_idata() f
[…]
