+++ pymc_marketing/mmm/__init__.py
@@ -14,6 +14,11 @@
 """Marketing Mix Models (MMM)."""
 
 from pymc_marketing.mmm import preprocessing, validating
+from pymc_marketing.mmm.budget_optimizer import (
+    BudgetOptimizer,
+    merge_inference_data,
+    merge_models_and_idata,
+)
 from pymc_marketing.mmm.components.adstock import (
     AdstockTransformation,
     BinomialAdstock,
@@ -84,6 +89,7 @@ __all__ = [
     "MMM",
     "AdstockTransformation",
     "BinomialAdstock",
+    "BudgetOptimizer",
     "BudgetOptimizerWrapper",
     "CovFunc",
     "DataDerivedScaling",
@@ -127,6 +133,8 @@ __all__ = [
     "create_constrained_inverse_gamma_prior",
     "create_eta_prior",
     "create_m_and_L_recommendations",
+    "merge_inference_data",
+    "merge_models_and_idata",
     "preprocessing",
     "preprocessing_method_X",
     "preprocessing_method_y",
