+++ pymc_marketing/mmm/budget_optimizer.py
@@ -192,27 +192,28 @@ Use a custom pymc model with any dimensionality
 Requirements
 ------------
 
-- The optimizer works on any wrapper that satisfies `OptimizerCompatibleModelWrapper`:
-
-  - Attributes: `adstock`, `_channel_scales`, `idata` (xr.DataTree with posterior)
-  - Method: `_set_predictors_for_optimization(num_periods) -> pm.Model` that returns a PyMC
-    model where a variable named `channel_data` exists with dims including `"date"` and all
-    budget dims (e.g., `("channel", "geo")`).
-    The optimizer replaces `channel_data` with the optimization variable under the hood.
-- Posterior must contain a response variable (default: `"total_media_contribution_original_scale"`)
-  or any custom `response_variable` you pass, and the required MMM deterministics (e.g.
-  `channel_contribution`).
-- For time distribution: pass a DataArray with dims `("date", *budget_dims)` and values along
-  `date` summing to 1 for each budget cell.
-- Bounds can be a dict only for single‑dimensional budgets; otherwise use an xarray.DataArray
-  (use `optimizer_xarray_builder(...)`).
+- Pass a ``pm.Model`` and ``xr.DataTree`` directly to :class:`BudgetOptimizer`.
+  The model must expose a ``channel_data`` variable (or whatever ``channel_data_var`` names)
+  with dims including ``"date"`` and all budget dims (e.g., ``("channel", "geo")``).
+  The optimizer replaces ``channel_data`` with the optimization variable under the hood.
+- Posterior must contain a response variable (default: ``"total_media_contribution_original_scale"``)
+  or any custom ``response_variable`` you pass, and the required MMM deterministics
+  (e.g. ``channel_contribution``).
+- For time distribution: pass a DataArray with dims ``("date", *budget_dims)`` and values along
+  ``date`` summing to 1 for each budget cell.
+- Bounds can be a dict only for single‑dimensional budgets; otherwise use an
+  xarray.DataArray (use ``optimizer_xarray_builder(...)``).
+- For backward compatibility, pass a legacy wrapper (implementing
+  :class:`OptimizerCompatibleModel`) as ``model=`` — the optimizer will unpack it
+  automatically.
 
 Notes
 -----
-- If `budgets_to_optimize` is not provided, the optimizer auto‑detects cells with historical
-  information using `idata.posterior["channel_contribution"].mean(("chain","draw","date")).astype(bool)`.
-- Default bounds are `[0, total_budget]` on each optimized cell.
-- Set `callback=True` in `allocate_budget(...)` to receive per‑iteration diagnostics
+- If ``budgets_to_optimize`` is not provided, the optimizer auto‑detects cells with
+  historical information using
+  ``idata.posterior.channel_contribution.mean(("chain","draw","date")).astype(bool)``.
+- Default bounds are ``[0, total_budget]`` on each optimized cell.
+- Set ``callback=True`` in ``allocate_budget(...)`` to receive per‑iteration diagnostics
   (objective, gradient, constraints) for monitoring.
 """
 
@@ -225,7 +226,14 @@ import pymc as pm
 import pytensor.tensor as pt
 import pytensor.xtensor as ptx
 import xarray as xr
-from pydantic import BaseModel, ConfigDict, Field, InstanceOf, PrivateAttr
+from pydantic import (
+    BaseModel,
+    ConfigDict,
+    Field,
+    InstanceOf,
+    PrivateAttr,
+    model_validator,
+)
 from pymc import Model, do
 from pymc.model.fgraph import clone_model
 from pymc.model.transform.optimization import freeze_dims_and_data
@@ -235,7 +243,7 @@ from pytensor.graph import rewrite_graph
 from pytensor.xtensor import as_xtensor
 from pytensor.xtensor.type import XTensorVariable
[…]

[…]
