### 2026-06-29 b84c46a
fix(docs): use inline code for internal OptimizerCompatibleModel reference
