### 2026-06-29 6b5d263
fix: address review feedback
- Add frozen_deterministics field to BudgetOptimizer, thread through
  extract_response_distribution and mmm.budget_optimizer()
- Gate mu_effects with NotImplementedError until #2621 lands
- Remove dead mu_effects scaffolding (_effect_budget_xtensors, slice code)
- Extract InferenceData->DataTree conversion into _to_datatree helper
- Remove redundant create_zero_dataset call in budget_optimizer
- Fix merge_models_and_idata docstring (remove stale 'at least 2' claim)
- Verify az.extract works with bare xr.Dataset
