### 2026-06-29 c01a309
[Notebook Re-Run] Bass + General Notebooks
- Bump arviz>=1.2.0, arviz-plots>=1.2.0 for arviz-darkgrid support
- Re-execute bass_example.ipynb with azp.plot_trace_dist
- Re-execute model_configuration.ipynb
- Re-execute other_nuts_samplers.ipynb
- Re-execute prior_predictive.ipynb
- Apply style migration: arviz-vibrant -> arviz-darkgrid

Related to #2614
