+++ pyproject.toml
@@ -25,8 +25,8 @@ classifiers = [
 ]
 
 dependencies = [
-    "arviz>=1.1.0,<2.0",
-    "arviz-plots[matplotlib]>=1.0.0,<2.0",
+    "arviz>=1.2.0,<2.0",
+    "arviz-plots[matplotlib]>=1.2.0,<2.0",
     "matplotlib>=3.5.1,<3.11",
     "narwhals",
     "numpy>=2.0,<3.0",
