### 2026-06-29 5245545
fix(docs): change Example: to Examples in extract_response_distribution docstring
