### 2026-06-29 ed8a289
Revert "fix(docs): suppress docutils 'Explicit markup' warning in Sphinx build"
This reverts commit 928d835664b16da3f1e25451844e9970b1d87061.
