### 2026-06-29 0f6666e
[Notebook Re-Run] MMM Fundamentals 1 (#2657)
* migration init

* rerun

* traceplot

* add titles and legends

* rm plot

* budget allocaton plot

* fine tune plots

* fix labelewr
