+++ .pre-commit-config.yaml
@@ -12,7 +12,7 @@ repos:
           - --exclude=scripts/
           - --start-year=2022
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.15.18
+    rev: v0.15.20
     hooks:
       - id: ruff-check
         types_or: [python, pyi, jupyter]
