### 2026-06-30 528bf79
add facets to channel contribution
