### 2026-06-30 3c5cbd9
Fix mypy errors on main (#2667)
Resolve the 6 mypy failures blocking pre-commit.ci on main:

- customer_choice/synthetic_data.py: cast the np.where index array to
  int so the loop variable and dict keys have a consistent int type
  (the variable was previously bound to int via an earlier range loop).
- mmm/hsgp.py: coerce the scalar operands to float before max()/int()
  so the function returns the annotated tuple[int, float].
- mmm/lift_test.py: widen the Index alias to also accept np.ndarray,
  matching np.argmax's return type assigned into indices[col].

Co-authored-by: Daimon <noreply@ant
[…]
