### 2026-06-30 6d6302f
Retry notebook execution on transient out-of-memory failures
The notebook CI (`Test Notebooks`) intermittently fails in the
`mmm --start-idx 10` split at `mmm_evaluation.ipynb` with
`MemoryError: std::bad_alloc`, raised by matplotlib while rendering a
figure (`RendererAgg.__init__`).

Investigation showed this is not a defect in the notebook:

- `mmm_evaluation.ipynb` runs fine standalone and within the full
  `mmm --start-idx 10` batch; peak memory across the whole batch is
  ~1.5 GB on an 8 GB box (the CI runner has 16 GB).
- The failing figure is an ordinary 1000x400 px plot. Rendering it 7
[…]
