### 2026-06-30 40ab831
DIAG: probe oversized matplotlib Agg allocations in notebook CI (temporary)
