### 2026-07-01 fb73e49
add mmm plot suite
