### 2026-07-01 5602e94
Make ci_bound coords explicit on whisker x; fix contribution_over_time hdi_prob docstring
Co-Authored-By: Claude Opus 4.8 (1M context) <noreply@anthropic.com>
