### 2026-07-01 63bf869
fix: suppress SettingWithCopyWarning and seaborn layout warnings in notebooks
TDD: added test_data_conversion.py with two red tests, then fixed
data_conversion.py:121 to use .loc instead of chained indexing.

Also removed conflicting constrained_layout rcParam in
mmm_geolift_calibration to silence seaborn axisgrid UserWarning.
