+++ tests/mmm/test_data_conversion.py
@@ -11,339 +11,68 @@
 #   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
-"""Tests for MMM data conversion utilities."""
+
+"""Tests for the data_conversion module."""
+
+import warnings
 
 import numpy as np
 import pandas as pd
 import pytest
-import xarray as xr
 
 from pymc_marketing.mmm.data_conversion import to_mmm_dataset
 
 
-class TestToMmmDatasetXInputTypes:
-    """to_mmm_dataset handles all supported X input types."""
+@pytest.fixture
+def sample_dataframe():
+    rng = np.random.default_rng(42)
+    n = 100
+    return pd.DataFrame(
+        {
+            "date": pd.date_range("2024-01-01", periods=n, freq="W-MON"),
+            "x1": rng.uniform(0.1, 1.0, n),
+            "x2": rng.uniform(0.1, 1.0, n),
+        }
+    )
 
-    def test_dataframe_x_with_series_y(self):
-        X = pd.DataFrame(
-            {
-                "date": pd.date_range("2023-01-01", periods=5, freq="W"),
-                "channel_1": [1.0, 2.0, 3.0, 4.0, 5.0],
-            }
-        )
-        y = pd.Series([10, 20, 30, 40, 50], name="target")
-        ds = to_mmm_dataset(X, y, date_column="date", channel_columns=["channel_1"])
-        assert "_channel" in ds.data_vars
-        assert "_target" in ds.data_vars
-        assert "date" in ds.coords
 
-    def test_x_as_dataarray(self):
-        da = xr.DataArray(
-            [[1, 2], [3, 4]],
-            dims=("date", "channel"),
-            coords={
-                "date": pd.date_range("2023-01-01", periods=2, freq="W"),
-                "channel": ["tv", "digital"],
-            },
-            name="media",
-        )
-        ds = to_mmm_dataset(da, date_column="date", channel_columns=["tv", "digital"])
-        assert "_channel" in ds.data_vars
-        assert ds.sizes["channel"] == 2
+def test_to_mmm_dataset_no_setting_with_copy_warning(sample_dataframe):
+    """to_mmm_dataset should not trigger SettingWithCopyWarning on DataFrame slices."""
+    df_slice = sample_dataframe[sample_dataframe["x1"] > 0.4]
 
-    def test_x_as_dataarray_without_name_defaults_to_media(self):
-        da = xr.DataArray(
-            [[1]],
-            dims=("date", "channel"),
-            coords={
-                "date": pd.date_range("2023-01-01", periods=1, freq="W"),
-                "channel": ["tv"],
-            },
-        )
-        ds = to_mmm_dataset(da, date_column="date", channel_columns=["tv"])
-        assert "_channel" in ds.data_vars
-
-    def test_x_as_dataset_renames_media_to_channel(self):
-        ds_in = xr.Dataset(
-            {"media": xr.DataArray([[1]], dims=("date", "channel"))},
-            coords={
-                "date": pd.date_range("2023-01-01", periods=1, freq="W"),
-                "channel": ["tv"],
-            },
-        )
-        ds_out = to_mmm_dataset(ds_in, date_column="date", channel_columns=["tv"])
-        assert "_channel" in ds_out.data_vars
-        assert "media" not in ds_out.data_vars
-
-    def test_x_as_dataset_renames_control_to_control(self):
-        ds_in = xr.Dataset(
-            {
-                "_channel": xr.DataArray([[1]], dims=("date", "channel")),
-                "control": xr.DataArray([1.0], dims=("date",)),
-            },
-            coords={
[…]
