### 2026-07-01 2548d53
chore: remove overkill data_conversion tests
The .loc fix in data_conversion.py is self-documenting best practice.
Testing pandas's own SettingWithCopyWarning is upstream territory.
