### 2026-07-01 99e1ac5
fix: preserve metadata attrs on sample_prior/sample_posterior_predictive return values
Pass keep_dataset=True to az.extract in both sample_prior_predictive and
sample_posterior_predictive to prevent xarray's Dataset.__getitem__ from
stripping group-level attrs when converting single-variable groups to DataArrays.
This restores metadata (created_at, inference_library, pymc_marketing_version, etc.)
on the returned Dataset, which was lost after the PyMC 6 / DataTree migration.
