### 2026-07-01 632dcdc
Add real attr assertions to prior/posterior predictive tests
Per review request: assert actual attr content rather than only isinstance.
- prior predictive: pymc_marketing_version equals __version__, and the
  standard ArviZ/PyMC metadata attrs (created_at, inference_library,
  inference_library_version) survive az.extract.
- posterior predictive: same metadata attrs are preserved through the
  extract conversion (the regression this PR fixes).

Co-Authored-By: Claude Opus 4.8 (1M context) <noreply@anthropic.com>
