### 2026-07-02 5e49925
Add merge_group trigger to CI workflows for merge queue support
