+++ .github/workflows/docs.yml
@@ -12,6 +12,7 @@ on:
     branches: [main, v1.0.0]
   push:
     branches: [main, v1.0.0]
+  merge_group:
 
 concurrency:
   group: ${{ github.workflow }}-${{ github.event.pull_request.number || github.ref }}
