### 2026-07-02 ab71419
Add credibility-theory campaign disaggregation demo
Standalone Quarto notebook (mmm_credibility_disaggregation.qmd) that splits an
MMM channel-level contribution across individual campaigns using Bühlmann
credibility (reach as the effective sample size), together with:
  - the six simulated data files it needs to run, and
  - simulate_campaign_data.py, the script that generates them (a user-level
    ad-world simulation plus a pymc-marketing MMM fit).

channel_contribution_posterior.nc is force-added because *.nc is gitignored.

Co-Authored-By: Claude Fable 5 <noreply@anthropic.com>
