+++ docs/source/notebooks/mmm/dev/data/simulate_campaign_data.py
@@ -1,0 +1,434 @@
+"""Simulate a synthetic social-media campaign world for MMM disaggregation research.
+
+Generative story (user-level, so reach/CTR dynamics emerge mechanically):
+
+1. A universe of ``N_USERS`` users, each belonging to one of ``K`` interest
+   groups and carrying an individual responsiveness score ``r_u``.
+2. Six campaigns on one "social" channel target overlapping subsets of the
+   groups -> audience sets with a known NxN overlap matrix.
+3. Every day, each active campaign buys impressions: ``spend / CPM * 1000``.
+   CPM inflates with audience penetration (auction depth). Impressions land on
+   the daily-online subset of the audience, weighted by responsiveness and a
+   mild retargeting bias -> cumulative reach follows a C-shaped saturating
+   curve with a knee, daily frequency keeps growing after the knee.
+4. Clicks: per-impression click prob = base CTR x fresh-user boost x
+   exp(-decay * prior frequency) -> CTR degrades once the pool saturates.
+5. True incremental conversions: exposures build per-user, per-campaign
+   "goodwill" (fresh exposures worth much more than repeats); goodwill decays
+   daily (user-level adstock). Daily conversion hazard = sum of goodwill;
+   realized conversions are attributed to campaigns by hazard share ->
+   GROUND-TRUTH campaign contributions.
+6. Sales y(t) = baseline trend + search-channel effect + value-per-conversion
+   x social conversions + noise.
+7. A pymc-marketing MMM (geometric adstock + logistic saturation) is fit on
+   (date, social spend, search spend, trend, y). The posterior of
+   original-scale channel contribution curves is saved -> this is "what the
+   model delivers" and what the disaggregation methods must distribute.
+
+Outputs (this folder):
+- campaign_cumulative_reach.csv   dates x campaigns, cumulative unique users
+- campaign_daily_metrics.csv      long: date, campaign, spend, impressions,
+                                  reach (daily uniques), clicks, active flag
+- audience_overlap_matrix.csv     NxN shared-audience counts (diag = size)
+- mmm_input_data.csv              date, social, search, trend, y
+- channel_contribution_posterior.nc  posterior draws of original-scale
+                                  contribution (chain, draw, date, channel)
+- ground_truth_campaign_contributions.csv  date x campaign true incremental
+                                  revenue (oracle - validation only)
+- ground_truth_channel_components.csv      date, baseline, search, social
+- ground_truth_meta.json          audiences, params, union reach, etc.
+
+Run:  uv run python simulate_campaign_data.py
+(The branch requires pymc>=6 / arviz>=1.2 - use the repo uv env, not stale
+conda envs.)
+"""
+
+import json
+from pathlib import Path
+
+import numpy as np
+import pandas as pd
+import xarray as xr
+
+SEED = 42
+N_USERS = 200_000
+N_DAYS = 120
+DATES = pd.date_range("2025-01-01", periods=N_DAYS, freq="D")
+GROUP_PROBS = [0.35, 0.30, 0.20, 0.15]
+P_ONLINE = 0.35  # fraction of audience in-auction on a given day
+P_CHURN = 0.012  # daily user turnover (new users / cookie resets) -> reach tail
+BURN_DAYS = 45  # warm-up so always-on campaigns start at steady state
+VALUE_PER_CONVERSION = 25.0
+OUT_DIR = Path(__file__).parent
+
+CAMPAIGNS = [
[…]
