+++ docs/source/notebooks/mmm/dev/mmm_credibility_disaggregation.qmd
@@ -1,0 +1,799 @@
+---
+title: "Who Earned the Credit?"
+subtitle: "Splitting a channel's Media-Mix impact across campaigns — using a 100-year-old idea from insurance"
+author: "PyMC-Marketing"
+date: last-modified
+format:
+  html:
+    toc: true
+    toc-depth: 2
+    number-sections: false
+    code-fold: true
+    code-tools: true
+    code-summary: "Show the code"
+    fig-align: center
+    theme: cosmo
+    embed-resources: true
+jupyter: python3
+execute:
+  warning: false
+---
+
+Your Media-Mix Model has done its job. It tells you, with honest uncertainty,
+that **Paid Social drove this much incremental revenue** — day by day. You bring
+it to the review, and the first question is the one the MMM can't answer:
+
+> *"Great — but **which campaign**?"*
+
+So you open the platform dashboard. Every campaign is proudly claiming
+conversions. Add them up and they total far more than the MMM says the whole
+channel earned. The always-on brand campaign looks steady; a flashy weekend
+burst looks like a superstar; a tiny niche campaign swings wildly from great to
+terrible week to week. And you already know the dirty secret: most of those
+"conversions" are rule-based and **non-incremental** — plenty of those people
+would have converted anyway.
+
+You have **one number you trust** (the MMM's channel total) and **six shaky
+signals** (the per-campaign platform numbers). You need to divide the trustworthy
+number across the campaigns. How much should you believe each campaign's own
+story?
+
+::: {.callout-tip title="The old idea we're going to borrow"}
+This is *exactly* the problem an insurance actuary solves before breakfast.
+
+Picture pricing car insurance. A driver comes in with **one** claim-free year.
+Do you give them the flawless-driver discount? Of course not — one year is too
+little to tell skill from luck. So you **blend** their record with the average
+driver, and you lean on their own history *more and more as it piles up*.
+
+That blend has a name — **credibility** — and a single dial, called **Z**, that
+runs from 0 ("I'll trust the average") to 1 ("I'll trust you"). The dial turns on
+one thing: **how much evidence you have.**
+:::
+
+**The whole idea of this notebook, in one sentence:** swap *"years of driving
+history"* for *"how many people the campaign reached,"* and use that dial to
+decide how much of each campaign's own story to believe.
+
+A campaign that reached 60,000 people has earned the right to be judged on its
+own numbers. One that reached 8,000 hasn't — its numbers are mostly noise, so we
+gently pull it toward the channel average until it earns our trust. **Reach is
+evidence.** Everything below is just that idea, done carefully and with honest
+uncertainty.
+
+::: {.callout-note collapse="true" title="For the statistically minded: the exact mapping (and why it equals Fay–Herriot)"}
+This is **Bühlmann credibility**, which blends a policyholder's own experience
+with the collective mean,
+$$\text{premium}=Z\cdot(\text{own})+(1-Z)\cdot(\text{collective}),\qquad Z=\frac{n}{n+k},$$
+and it is *exactly* hierarchical-Bayes partial pooling.
+
+| Bühlmann credibility (actuarial) | This MMM disaggregation |
+|---|---|
+| Policyholder | Campaign $c$ |
+| Portfolio / *collective* | The whole social channel |
+| Own loss experience (noisy) | Naive contribution-per-reached-user (social split by click share) |
[…]
