### 2026-07-02 19cab3c
Add April broad_prospecting campaign that never saturates
New seventh campaign kicks off April 1st against a very large audience,
so its expanding-window reach is still growing when the window closes.
Regenerate all simulated data + MMM posterior and update the notebook
narrative: the newcomer carries the least evidence, gets the lowest Z,
and is shrunk hardest toward the channel mean. Include rendered HTML.
