+++ docs/source/notebooks/mmm/dev/mmm_credibility_disaggregation.qmd
@@ -28,15 +28,16 @@ it to the review, and the first question is the one the MMM can't answer:
 So you open the platform dashboard. Every campaign is proudly claiming
 conversions. Add them up and they total far more than the MMM says the whole
 channel earned. The always-on brand campaign looks steady; a flashy weekend
-burst looks like a superstar; a tiny niche campaign swings wildly from great to
-terrible week to week. And you already know the dirty secret: most of those
-"conversions" are rule-based and **non-incremental** — plenty of those people
-would have converted anyway.
+burst looks like a superstar; a tiny niche campaign swings from hero to flop
+week to week; and the broad prospecting push that only launched in April has
+barely four weeks of history to its name. And you already know the dirty
+secret: most of those "conversions" are rule-based and **non-incremental** —
+plenty of those people would have converted anyway.
 
-You have **one number you trust** (the MMM's channel total) and **six shaky
-signals** (the per-campaign platform numbers). You need to divide the trustworthy
-number across the campaigns. How much should you believe each campaign's own
-story?
+Strip the noise away and the problem is simple. You have **one number you
+trust** (the MMM's channel total) and **seven shaky signals** (the per-campaign
+platform numbers). The question is not *whether* to use the shaky signals —
+they're all you have — it's **how much to believe each one**.
 
 ::: {.callout-tip title="The old idea we're going to borrow"}
 This is *exactly* the problem an insurance actuary solves before breakfast.
@@ -44,22 +45,33 @@ This is *exactly* the problem an insurance actuary solves before breakfast.
 Picture pricing car insurance. A driver comes in with **one** claim-free year.
 Do you give them the flawless-driver discount? Of course not — one year is too
 little to tell skill from luck. So you **blend** their record with the average
-driver, and you lean on their own history *more and more as it piles up*.
+driver, and you lean on their own history *more and more as it piles up*:
 
-That blend has a name — **credibility** — and a single dial, called **Z**, that
-runs from 0 ("I'll trust the average") to 1 ("I'll trust you"). The dial turns on
-one thing: **how much evidence you have.**
+$$\text{best guess}=Z\cdot(\text{their record})+(1-Z)\cdot(\text{average driver}),
+\qquad Z=\frac{n}{n+k}.$$
+
+That blend has a name — **credibility** — and a single dial, **Z**, that runs
+from 0 ("I'll trust the average") to 1 ("I'll trust you"). It turns on one thing:
+**how much evidence you have** ($n$), measured against a threshold $k$ that says
+how much evidence counts as "enough."
+
+The arithmetic is friendly. Say $k$ works out to 20,000 reached people. A
+campaign that reached 60,000 gets $Z=\tfrac{60}{60+20}=0.75$ — we believe 75% of
+its own story and fill the rest with the group average. A campaign that reached
+8,000 gets $Z=\tfrac{8}{8+20}\approx 0.29$ — mostly the average, until it earns
+more trust. Nobody is ignored; nobody is blindly believed.
 :::
 
 **The whole idea of this notebook, in one sentence:** swap *"years of driving
 history"* for *"how many people the campaign reached,"* and use that dial to
-decide how much of each campaign's own story to believe.
+decide how much of each campaign's own story to believe. **Reach is evidence.**
+Everything below is just that idea, done carefully and with honest uncertainty.
 
-A campaign that reached 60,000 people has earned the right to be judged on its
-own numbers. One that reached 8,000 hasn't — its numbers are mostly noise, so we
-gently pull it toward the channel average until it earns our trust. **Reach is
-evidence.** Everything below is just that idea, done carefully and with honest
-uncertainty.
[…]
