### 2026-07-02 89243b6
Convert credibility disaggregation demo to Jupyter notebook
Replace the Quarto source (.qmd) and rendered HTML with an executed
notebook. The simulation script and data files are unchanged.
