### 2026-07-03 4ae3fcc
fix(docs): use 'docutils' in suppress_warnings to match warning source tag
