### 2026-07-04 16fa7bd
fix(plot): use matplotlib date converter for datetime x-axis
plot_curve drew sample lines with pandas period ordinals while the HDI
band used matplotlib date ordinals, producing broken x-axis tick labels
on datetime coordinates. Pass x_compat=True so pandas uses the
matplotlib date converter. No-op for numeric x-axes.

Fixes #2682
