### 2026-07-05 6e3bf8f
test(mlflow): force pymc sampler for bass autolog metrics on v1
Under pymc6 the default sampler (nutpie) does not log sampling_time /
time_per_draw, so the strict metrics assertion in test_autolog_bass fails.
Force nuts_sampler=pymc, matching the v1 MMM/CLV autolog tests.
