### 2026-07-05 26b33b1
fix(bass): store deterministics on posterior DataTree node
Under arviz>=1.2 InferenceData subclasses xarray.DataTree, so the
idata.posterior = compute_deterministics(...) assignment in BassModel.fit
sets a shadowing instance attribute and leaves the group unchanged. The
deterministics (adopters, innovators, imitators, peak) were dropped from
the fitted posterior. Assign to the idata["posterior"] node instead.
