### 2026-07-05 4c458b4
refactor(bass): migrate plot_peak to arviz_plots.plot_dist
arviz.plot_posterior was removed in arviz 1.2. Use arviz_plots.plot_dist
and extract the figure and axes from the returned PlotCollection, faceting
by product for multi-product models.
