### 2026-07-06 a7f150a
test(bass): guard the deterministics fix via the posterior accessor
test_fit_deterministics asserted against idata.posterior, which under
arviz>=1.2 returns a shadowing attribute and passes even when the
deterministics are missing from the DataTree node. Assert against the
public model.posterior accessor instead, so the test fails if fit stops
storing the deterministics on the node.
