### 2026-07-06 5d10115
Update data before repeated fit
