### 2026-07-06 ee90470
refactor(bass): migrate dim handling to pymc.dims
Replace create_dim_handler (pymc_extras.prior) with pymc.dims, closing
the last major create_dim_handler usage outside the Prior class (#2598).

- m/p/q created with xdist=True; deterministics via pmd.Deterministic
- inline xtensor _f_xt/_F_xt for the model graph; public F/f stay on
  PyTensor for the float/NumPy API and tests
- keep t and y_obs as pm.Data wrapped with pmd.as_xtensor so the
  data-setter and out-of-sample forecasting keep working
- transpose deterministics back to the historical (T, ...) dim order
- DiracDelta is not yet in pym
[…]
