### 2026-07-08 dc16806
remove t_unobserved plotting param
