### 2026-07-08 09eb485
remove pnbd fit_method param
