### 2026-07-08 94259c6
remove fit_method param
