### 2026-07-08 02f8208
remove data param from CLVModel
