### 2026-07-08 40fcb12
WIP remove data param from all models
