+++ pymc_marketing/clv/models/shifted_beta_geo.py
@@ -173,13 +173,11 @@ class ShiftedBetaGeoModel(CLVModel):
 
     def __init__(
         self,
-        data: pd.DataFrame | None = None,
         *,
         model_config: ModelConfig | None = None,
         sampler_config: dict | None = None,
     ):
         super().__init__(
-            data=data,
             model_config=model_config,
             sampler_config=sampler_config,
             non_distributions=["dropout_covariate_cols"],
