+++ pymc_marketing/clv/models/beta_geo.py
@@ -153,13 +153,11 @@ class BetaGeoModel(CLVModel):
 
     def __init__(
         self,
-        data: pd.DataFrame | None = None,
         *,
         model_config: dict | None = None,
         sampler_config: dict | None = None,
     ):
         super().__init__(
-            data=data,
             model_config=model_config,
             sampler_config=sampler_config,
             non_distributions=["purchase_covariate_cols", "dropout_covariate_cols"],
