+++ pymc_marketing/clv/models/beta_geo_beta_binom.py
@@ -147,13 +147,11 @@ class BetaGeoBetaBinomModel(CLVModel):
 
     def __init__(
         self,
-        data: pd.DataFrame | None = None,
         *,
         model_config: ModelConfig | None = None,
         sampler_config: dict | None = None,
     ):
         super().__init__(
-            data=data,
             model_config=model_config,
             sampler_config=sampler_config,
             non_distributions=None,
