### 2026-07-08 45fdac9
remove old API  idata_to_init_kwargs
