### 2026-07-09 8792958
CLVModel fit logic
