### 2026-07-09 e8f6a48
PPC plotting logic for unfit models
