### 2026-07-09 9b1489a
BetaGeoModel build_model logic
