### 2026-07-09 f176604
ParetoNBDModel build_model logic and tests
