### 2026-07-09 a4aa988
BGBB and GGM build_model_logic
