+++ pymc_marketing/clv/models/gamma_gamma.py
@@ -331,28 +331,16 @@ class GammaGammaModel(BaseGammaGammaModel):
             must_be_unique=["customer_id"],
         )
 
-    def build_model(self, data: pandas.DataFrame | None = None) -> None:  # type: ignore[override]
+    def build_model(self, data: pandas.DataFrame) -> None:  # type: ignore[override]
         """Build the model.
 
         Parameters
         ----------
         data : pandas.DataFrame, optional
             Input data with customer_id, monetary_value, and frequency columns.
-            If not provided, uses data from model initialization (deprecated).
         """
-        # TODO: Revise this logic when old API is removed in 1.0.
-        # Handle data parameter
-        if data is not None:
-            self._validate_data(data)
-            self.data = data
-        elif not hasattr(self, "data") or self.data is None:
-            raise ValueError(
-                f"{self._model_type}.build_model() requires data parameter. "
-                "Either pass data to build_model(data=...) or fit(data=...)"
-            )
-        else:
-            # Validate existing data from old API
-            self._validate_data(self.data)
+        self._validate_data(data)
+        self.data = data
 
         z_mean = pt.as_tensor_variable(self.data["monetary_value"])
         x = pt.as_tensor_variable(self.data["frequency"])
@@ -490,28 +478,16 @@ class GammaGammaModelIndividual(BaseGammaGammaModel):
             data, required_cols=["customer_id", "individual_transaction_value"]
         )
 
-    def build_model(self, data: pandas.DataFrame | None = None) -> None:  # type: ignore[override]
+    def build_model(self, data: pandas.DataFrame) -> None:  # type: ignore[override]
         """Build the model.
 
         Parameters
         ----------
         data : pandas.DataFrame, optional
             Input data with customer_id and individual_transaction_value columns.
-            If not provided, uses data from model initialization (deprecated).
         """
-        # TODO: Revise this logic when old API is removed in 1.0.
-        # Handle data parameter
-        if data is not None:
-            self._validate_data(data)
-            self.data = data
-        elif not hasattr(self, "data") or self.data is None:
-            raise ValueError(
-                f"{self._model_type}.build_model() requires data parameter. "
-                "Either pass data to build_model(data=...) or fit(data=...)"
-            )
-        else:
-            # Validate existing data from old API
-            self._validate_data(self.data)
+        self._validate_data(data)
+        self.data = data
 
         z = self.data["individual_transaction_value"]
 
