### 2026-07-09 1afb978
GGM tests, MBG logic, build_model docstrings
