+++ pymc_marketing/clv/models/pareto_nbd.py
@@ -256,7 +256,7 @@ class ParetoNBDModel(CLVModel):
 
         Parameters
         ----------
-        data : pd.DataFrame, optional
+        data : pd.DataFrame
             Input data with customer_id, frequency, recency, and T columns.
         """
         self._validate_data(data)
