+++ pymc_marketing/clv/models/beta_geo.py
@@ -212,7 +212,7 @@ class BetaGeoModel(CLVModel):
 
         Parameters
         ----------
-        data : pd.DataFrame, optional
+        data : pd.DataFrame
             Input data with customer_id, frequency, recency, and T columns.
         """
         self._validate_data(data)
