### 2026-07-09 ed0c49a
mbg tests update
