### 2026-07-09 c594fbe
sBG build logic and unit test updates
