+++ tests/clv/models/test_shifted_beta_geo.py
@@ -48,7 +48,6 @@ class TestShiftedBetaGeoModel:
         cls.data = pd.read_csv("data/sbg_cohorts.csv").query("T <= 8")
         cls.model = ShiftedBetaGeoModel()
         cls.model.build_model(data=cls.data)
-        cls.model.build_model()
 
         # Mock an idata object for tests requiring a fitted model
         cls.N = len(cls.data)
@@ -203,16 +202,13 @@ class TestShiftedBetaGeoModel:
         }
 
     def test_model(self, custom_model_config):
-        default_model = ShiftedBetaGeoModel(
-            data=self.data,
-        )
+        default_model = ShiftedBetaGeoModel()
         custom_model = ShiftedBetaGeoModel(
-            data=self.data,
             model_config=custom_model_config,
         )
 
         for model in (default_model, custom_model):
-            model.build_model()
+            model.build_model(data=self.data)
             assert isinstance(
                 model.model["alpha"].owner.op,
                 ViewOp | Elemwise
@@ -348,11 +344,8 @@ class TestShiftedBetaGeoModel:
             ValueError,
             match=r"The following required columns are missing from the input data: \['customer_id'\]",
         ):
-            with pytest.warns(
-                DeprecationWarning, match="will be removed in version 1.0"
-            ):
-                model = ShiftedBetaGeoModel(data=data_invalid)
-            model.build_model()
+            model = ShiftedBetaGeoModel()
+            model.build_model(data=data_invalid)
 
         data_invalid = self.data.drop(columns="recency")
 
@@ -360,11 +353,8 @@ class TestShiftedBetaGeoModel:
             ValueError,
             match=r"The following required columns are missing from the input data: \['recency'\]",
         ):
-            with pytest.warns(
-                DeprecationWarning, match="will be removed in version 1.0"
-            ):
-                model = ShiftedBetaGeoModel(data=data_invalid)
-            model.build_model()
+            model = ShiftedBetaGeoModel()
+            model.build_model(data=data_invalid)
 
         data_invalid = self.data.drop(columns="T")
 
@@ -372,11 +362,8 @@ class TestShiftedBetaGeoModel:
             ValueError,
             match=r"The following required columns are missing from the input data: \['T'\]",
         ):
-            with pytest.warns(
-                DeprecationWarning, match="will be removed in version 1.0"
-            ):
-                model = ShiftedBetaGeoModel(data=data_invalid)
-            model.build_model()
+            model = ShiftedBetaGeoModel()
+            model.build_model(data=data_invalid)
 
     def test_customer_id_duplicate(self):
         with pytest.raises(
@@ -390,11 +377,8 @@ class TestShiftedBetaGeoModel:
                     "cohort": np.asarray(["A", "A"]),
                 }
             )
-            with pytest.warns(
-                DeprecationWarning, match="will be removed in version 1.0"
-            ):
-                model = ShiftedBetaGeoModel(data=data)
-            model.build_model()
+            model = ShiftedBetaGeoModel()
+            model.build_model(data=data)
 
     def test_invalid_recency(self):
         data = pd.DataFrame(
@@ -408,11 +392,8 @@ class TestShiftedBetaGeoModel:
         with pytest.raises(
             ValueError, match=r"Model fitting requires 1 <= recency <= T, and T >= 2."
         ):
-            with pytest.warns(
-                DeprecationWarning, match="will be removed in version 1.0"
-            ):
-                model = ShiftedBetaGeoModel(data=data)
-            model.build_model()
+            model = ShiftedBetaGeoModel()
+            model.build_model(data=data)
 
     def test_invalid_T(self):
[…]
