### 2026-07-09 1f5cf67
remove deprecated test
