### 2026-07-09 6b77356
remove fit_method param from mlflow.py
