### 2026-07-09 19a74a6
update runslow bg and mbg tests
