+++ tests/clv/models/test_modified_beta_geo.py
@@ -243,7 +243,7 @@ class TestModifiedBetaGeoModel:
             if method == "mcmc"
             else {}
         )
-        model.fit(method=method, progressbar=False, **sample_kwargs)
+        model.fit(data=self.data, method=method, progressbar=False, **sample_kwargs)
 
         fit = model.idata.posterior
         np.testing.assert_allclose(
