### 2026-07-09 fd3f14b
remove default data arg in mlflow patch_clv_fit
