+++ pymc_marketing/mlflow.py
@@ -1347,7 +1347,7 @@ def autolog(
 
     def patch_clv_fit(fit):
         @wraps(fit)
-        def new_fit(self, data=None, method: str = "mcmc", **kwargs):
+        def new_fit(self, data, method: str = "mcmc", **kwargs):
             mlflow.log_param("model_type", self._model_type)
             mlflow.log_param(
                 "fit_method",
