### 2026-07-09 3f639e1
refactor(bass): address review on the pymc.dims migration
- consolidate F/f: a single _exp helper dispatches pmd.math.exp for xtensor
  and pt.exp otherwise, so F/f serve both the float/NumPy API and the model
  graph; drop the duplicate _f_xt/_F_xt (williambdean)
- drop the _ordered transpose; let pymc.dims order the deterministic dims,
  consumers select by name (williambdean)
