### 2026-07-09 dc9e6f6
fix(bass): move plot_decomposition legend outside the axes
Place the legend as a figure legend to the right of the grid instead of
inside the first subplot, where it overlapped the curves (williambdean).
