### 2026-07-10 954308d
fix(bass): put plot_decomposition legend below the grid
A horizontal legend centered under the subplots reads better than the
right-side placement, which sat next to the empty facet cell (williambdean).
