### 2026-07-10 6f48be5
feat(mlflow): autolog support for BassModel (#2624)
* feat(mlflow): autolog support for BassModel

* fix(mlflow): keep autolog running when model graph creation fails

* docs(mlflow): fix log_bass_configuration docstring and Bass examples

* test(mlflow): force pymc sampler for bass autolog metrics on v1

Under pymc6 the default sampler (nutpie) does not log sampling_time /
time_per_draw, so the strict metrics assertion in test_autolog_bass fails.
Force nuts_sampler=pymc, matching the v1 MMM/CLV autolog tests.
