### 2026-07-10 e32e536
refactor(bass): generalize the plotting facet dimension
The plot methods hardcoded a 'product' dimension. Infer the faceting
dimension from the data (the non-time, non-sample dim, the same one
plot_curve facets over), with an optional dim override and a coord
selector. So a model whose series dimension is named e.g. 'region'
plots and selects correctly with no extra arguments. Adds region-model
tests, including an observed-overlay guard (williambdean).
