+++ pymc_marketing/bass/plotting.py
@@ -15,8 +15,9 @@
 
 Each function takes a fitted :class:`~pymc_marketing.bass.model.BassModel`
 and returns a ``(Figure, ndarray of Axes)`` tuple, following the convention
-of :func:`pymc_marketing.plot.plot_curve`. Multi-product models plot one
-subplot per product; pass ``product`` to select a single one.
+of :func:`pymc_marketing.plot.plot_curve`. Multi-series models plot one
+subplot per coordinate of the faceting dimension (inferred from the data,
+or set with ``dim``); pass ``coord`` to select a single one.
 
 The functions are also exposed as ``plot_*`` methods on
 :class:`~pymc_marketing.bass.model.BassModel`.
@@ -45,37 +46,51 @@ __all__ = [
 ]
 
 
-def _select_product(da: xr.DataArray, product: str | None) -> xr.DataArray:
-    if product is None:
+def _facet_dim(da: xr.DataArray, dim: str | None) -> str | None:
+    """Return the dimension the plots facet over.
+
+    ``dim`` if given, otherwise the single non-time, non-sample
+    dimension of ``da`` (the same one :func:`plot_curve` facets over), or
+    ``None`` for a single-series model.
+    """
+    if dim is not None:
+        return dim
+    extra = [d for d in da.dims if d not in ("chain", "draw", "T")]
+    return extra[0] if extra else None
+
+
+def _select(da: xr.DataArray, dim: str | None, coord: str | None) -> xr.DataArray:
+    if coord is None:
         return da
-    if "product" not in da.dims:
+    if dim is None or dim not in da.dims:
         raise ValueError(
-            "The model has no 'product' dimension. "
-            "Remove the product argument for single-product models."
+            "The model has no faceting dimension. "
+            "Remove the coord argument for single-series models."
         )
-    return da.sel(product=product)
+    return da.sel({dim: coord})
 
 
-def _get_observed(model: "BassModel", product: str | None) -> xr.DataArray | None:
+def _get_observed(
+    model: "BassModel", dim: str | None, coord: str | None
+) -> xr.DataArray | None:
     idata = model.idata
     if idata is None or "fit_data" not in idata or "observed" not in idata.fit_data:
         return None
-    return _select_product(idata.fit_data["observed"], product)
+    return _select(idata.fit_data["observed"], dim, coord)
 
 
 def _overlay_observed(
     observed: xr.DataArray,
     axes: npt.NDArray[Axes],
+    dim: str | None,
     **plot_kwargs: Any,
 ) -> None:
-    """Plot observed data on each axes, one product per axes."""
+    """Plot observed data on each axes, one coordinate per axes."""
     plot_kwargs = {"color": "black", **plot_kwargs}
     t = observed.coords["T"].values
-    if "product" in observed.dims:
-        for ax, product in zip(
-            axes.flat, observed.coords["product"].values, strict=False
-        ):
-            ax.plot(t, observed.sel(product=product).values, **plot_kwargs)
+    if dim is not None and dim in observed.dims:
+        for ax, coord in zip(axes.flat, observed.coords[dim].values, strict=False):
+            ax.plot(t, observed.sel({dim: coord}).values, **plot_kwargs)
     else:
         for ax in axes.flat:
             ax.plot(t, observed.values, **plot_kwargs)
@@ -83,7 +98,8 @@ def _overlay_observed(
 
 def plot_adoption_curve(
     model: "BassModel",
-    product: str | None = None,
+    dim: str | None = None,
+    coord: str | None = None,
     n_samples: int = 10,
     hdi_probs: float | list[float] | None = None,
     random_seed: np.random.Generator | None = None,
@@ -99,9 +115,12 @@ def plot_adoption_curve(
     ----------
[…]
