### 2026-07-12 e9ff8eb
unit test compat=override all CLV models
