+++ tests/clv/models/test_pareto_nbd.py
@@ -249,6 +249,21 @@ class TestParetoNBDModel:
             "\nrecency_frequency~ParetoNBD(r,alpha,s,beta,<constant>)"
         )
 
+    def test_extract_predictive_variables_no_future_warning(self):
+        import warnings
+
+        pred_data = self.data.iloc[:25].assign(future_t=1)
+        with warnings.catch_warnings():
+            warnings.simplefilter("error", FutureWarning)
+            dataset = self.model._extract_predictive_variables(
+                pred_data,
+                customer_varnames=["frequency", "recency", "T", "future_t"],
+            )
+        assert "r" in dataset
+        assert "alpha" in dataset
+        assert "s" in dataset
+        assert "beta" in dataset
+
     @pytest.mark.parametrize("future_t", [1, 3, 6])
     def test_expected_purchases(self, future_t):
         # Reference values from Pareto/NBD MLE on CDNOW subsample (25 customers)
