+++ tests/clv/models/test_beta_geo.py
@@ -557,6 +557,20 @@ class TestBetaGeoModel:
             rtol=0.001,
         )
 
+    def test_extract_predictive_variables_no_future_warning(self):
+        import warnings
+
+        with warnings.catch_warnings():
+            warnings.simplefilter("error", FutureWarning)
+            dataset = self.model._extract_predictive_variables(
+                self.pred_data.assign(future_t=1),
+                customer_varnames=["frequency", "recency", "T", "future_t"],
+            )
+        assert "a" in dataset
+        assert "b" in dataset
+        assert "alpha" in dataset
+        assert "r" in dataset
+
     @pytest.mark.parametrize("test_t", [1, 3, 6])
     def test_expected_purchases_new_customer(self, test_t):
         # Reference scalar values from BG/NBD MLE model
