+++ pymc_marketing/clv/models/beta_geo.py
@@ -467,7 +467,8 @@ class BetaGeoModel(CLVModel):
                 alpha,
                 r,
                 *customer_vars,
-            )
+            ),
+            compat="override",
         )
 
     def expected_purchases(
