+++ pymc_marketing/clv/models/shifted_beta_geo.py
@@ -538,7 +538,8 @@ class ShiftedBetaGeoModel(CLVModel):
                 alpha_pred,
                 beta_pred,
                 *customer_vars,
-            )
+            ),
+            compat="override",
         ).swap_dims(
             {"customer_id": "cohort"}
         )  # swap dims to enable cohort selection for predictions
