### 2026-07-13 a53561d
change default ggm priors to Weibull
