### 2026-07-13 29fcddc
Remove support for pre-v1.0.0 CLV API (#2723)
* remove t_unobserved plotting param

* remove pnbd fit_method param

* remove fit_method param

* remove data param from CLVModel

* WIP remove data param from all models

* remove data param from all models

* add uv.lock file to gitignore

* remove old API  idata_to_init_kwargs

* CLVModel fit logic

* PPC plotting logic for unfit models

* BetaGeoModel build_model logic

* ParetoNBDModel build_model logic and tests

* BGBB and GGM build_model_logic

* GGM tests, MBG logic, build_model docstrings

* mbg tests update

* sBG build logic and unit t
[…]
