+++ pymc_marketing/clv/models/pareto_nbd.py
@@ -197,13 +197,11 @@ class ParetoNBDModel(CLVModel):
 
     def __init__(
         self,
-        data: pd.DataFrame | None = None,
         *,
         model_config: ModelConfig | None = None,
         sampler_config: dict | None = None,
     ):
         super().__init__(
-            data=data,
             model_config=model_config,
             sampler_config=sampler_config,
             non_distributions=["purchase_covariate_cols", "dropout_covariate_cols"],
@@ -253,28 +251,16 @@ class ParetoNBDModel(CLVModel):
             must_be_unique=["customer_id"],
         )
 
-    def build_model(self, data: pd.DataFrame | None = None) -> None:  # type: ignore[override]
+    def build_model(self, data: pd.DataFrame) -> None:  # type: ignore[override]
         """Build the model.
 
         Parameters
         ----------
-        data : pd.DataFrame, optional
+        data : pd.DataFrame
             Input data with customer_id, frequency, recency, and T columns.
-            If not provided, uses data from model initialization (deprecated).
         """
-        # TODO: Revise this logic when old API is removed in 1.0.
-        # Handle data parameter
-        if data is not None:
-            self._validate_data(data)
-            self.data = data
-        elif not hasattr(self, "data") or self.data is None:
-            raise ValueError(
-                f"{self._model_type}.build_model() requires data parameter. "
-                "Either pass data to build_model(data=...) or fit(data=...)"
-            )
-        else:
-            # Validate existing data from old API
-            self._validate_data(self.data)
+        self._validate_data(data)
+        self.data = data
 
         coords = {
             "purchase_covariate": self.purchase_covariate_cols,
@@ -352,9 +338,8 @@ class ParetoNBDModel(CLVModel):
 
     def fit(  # type: ignore[override]
         self,
-        data: pd.DataFrame | None = None,
+        data: pd.DataFrame,
         method: str = "map",
-        fit_method: str | None = None,
         **kwargs,
     ):  # type: ignore
         """Infer posteriors of model parameters to run predictions.
@@ -373,15 +358,6 @@ class ParetoNBDModel(CLVModel):
 
         """
         mode = get_default_mode()
-
-        if fit_method:
-            warnings.warn(
-                "'fit_method' is deprecated and will be removed in version 1.0. "
-                "Use 'method' instead.",
-                DeprecationWarning,
-                stacklevel=2,
-            )
-            method = fit_method
 
         if method == "mcmc":
             # Include rewrite in mode
