### 2026-07-13 b84aab2
pytest filterwarning refactor
