+++ tests/clv/models/test_pareto_nbd.py
@@ -249,16 +249,13 @@ class TestParetoNBDModel:
             "\nrecency_frequency~ParetoNBD(r,alpha,s,beta,<constant>)"
         )
 
+    @pytest.mark.filterwarnings("error::FutureWarning")
     def test_extract_predictive_variables_no_future_warning(self):
-        import warnings
-
         pred_data = self.data.iloc[:25].assign(future_t=1)
-        with warnings.catch_warnings():
-            warnings.simplefilter("error", FutureWarning)
-            dataset = self.model._extract_predictive_variables(
-                pred_data,
-                customer_varnames=["frequency", "recency", "T", "future_t"],
-            )
+        dataset = self.model._extract_predictive_variables(
+            pred_data,
+            customer_varnames=["frequency", "recency", "T", "future_t"],
+        )
         assert "r" in dataset
         assert "alpha" in dataset
         assert "s" in dataset
