### 2026-07-13 83c2d27
pytest futurewarning regression testing
