### 2026-07-14 ff70aae
Fix `xarray` combine_by_coords FutureWarning (#2728)
* unit test compat=override all CLV models

* pytest filterwarning refactor

* rerun sbg notebook to clear warnings

* pytest futurewarning regression testing
