+++ pymc_marketing/clv/models/beta_geo.py
@@ -444,7 +444,8 @@ class BetaGeoModel(CLVModel):
                 alpha,
                 r,
                 *customer_vars,
-            )
+            ),
+            compat="override",
         )
 
     def expected_purchases(
