+++ pymc_marketing/clv/models/pareto_nbd.py
@@ -482,7 +482,8 @@ class ParetoNBDModel(CLVModel):
                 s,
                 beta,
                 *customer_vars,
-            )
+            ),
+            compat="override",
         )
 
     def expected_purchases(
