+++ tests/clv/models/test_beta_geo_beta_binom.py
@@ -25,6 +25,8 @@ from pymc_marketing.clv.distributions import BetaGeoBetaBinom
 from pymc_marketing.clv.models import BetaGeoBetaBinomModel
 from tests.clv.conftest import create_mock_fit, mock_sample
 
+pytestmark = pytest.mark.filterwarnings("error::FutureWarning")
+
 
 class TestBetaGeoBetaBinomModel:
     @classmethod
