### 2026-07-14 7af611a
remove ShiftedBetaGeoModelIndividual
