+++ tests/clv/models/test_shifted_beta_geo.py
@@ -11,19 +11,15 @@
 #   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
-import arviz as az
 import numpy as np
 import pandas as pd
-import pymc as pm
 import pytest
 import xarray as xr
-from pymc.distributions.censored import CensoredRV
 from pymc_extras.prior import Prior
 from pytensor.compile import ViewOp
 from pytensor.tensor.elemwise import Elemwise
-from scipy import stats
 
-from pymc_marketing.clv import ShiftedBetaGeoModel, ShiftedBetaGeoModelIndividual
+from pymc_marketing.clv import ShiftedBetaGeoModel
 from tests.clv.conftest import mock_sample
 
 
@@ -1124,388 +1120,3 @@ class TestShiftedBetaGeoModel:
         # Verify alpha has proper dimensions
         assert "alpha" in dataset
         assert dataset["alpha"].dims[-1] == "cohort"  # After swap_dims
-
-
-class TestShiftedBetaGeoModelIndividual:
-    @classmethod
-    def setup_class(cls):
-        def churned_data_from_percentage_alive(percentage_alive, initial_customers):
-            n_alive = (np.asarray(percentage_alive) / 100 * initial_customers).astype(
-                int
-            )
-
-            churned_at = np.zeros((initial_customers,), dtype=int)
-            counter = 0
-            for t, diff in enumerate((n_alive[:-1] - n_alive[1:]), start=1):
-                churned_at[counter : counter + diff] = t
-                counter += diff
-
-            censoring_t = t + 1
-            churned_at[counter:] = censoring_t
-
-            return churned_at
-
-        # Regular dataset from Fader, P. S., & Hardie, B. G. (2007). How to project customer retention.
-        # Journal of Interactive Marketing, 21(1), 76-90. https://journals.sagepub.com/doi/pdf/10.1002/dir.20074
-        cls.N = 1000
-        cls.T = 8
-        cls.customer_id = np.arange(cls.N)
-        cls.churn_time = churned_data_from_percentage_alive(
-            percentage_alive=[100.0, 63.1, 46.8, 38.2, 32.6, 28.9, 26.2, 24.1],
-            initial_customers=cls.N,
-        )
-        cls.ref_MLE_estimates = {"alpha": 0.688, "beta": 1.182}
-
-    @pytest.fixture(scope="class")
-    def model_config(self):
-        return {
-            "alpha": Prior("HalfNormal", sigma=10),
-            "beta": Prior("HalfStudentT", nu=4, sigma=10),
-        }
-
-    @pytest.fixture(scope="class")
-    def default_model_config(self):
-        return {
-            "alpha": Prior("HalfFlat"),
-            "beta": Prior("HalfFlat"),
-        }
-
-    @pytest.fixture(scope="class")
-    def data(self):
-        return pd.DataFrame(
-            {
-                "customer_id": self.customer_id,
-                "t_churn": self.churn_time,
-                "T": self.T,
-            }
-        )
-
-    def test_missing_cols(self, data):
-        # Create a version of the data that's missing the 'customer_id' column
-        data_invalid = data.drop(columns="customer_id")
-
-        with pytest.raises(
-            ValueError,
-            match=r"The following required columns are missing from the input data: \['customer_id'\]",
-        ):
-            with pytest.warns(
-                DeprecationWarning, match="will be removed in version 1.0"
-            ):
-                model = ShiftedBetaGeoModelIndividual(data=data_invalid)
-            model.build_model()
-
-        data_invalid = data.drop(columns="t_churn")
-
-        with pytest.raises(
-            ValueError,
[…]
