+++ pymc_marketing/clv/models/shifted_beta_geo.py
@@ -19,7 +19,6 @@ import numpy as np
 import pandas as pd
 import pymc as pm
 import xarray
-from pymc.util import RandomState
 from pymc_extras.prior import Prior
 from scipy.special import gammaln, hyp2f1
 
@@ -28,7 +27,7 @@ from pymc_marketing.clv.models import CLVModel
 from pymc_marketing.clv.utils import to_xarray
 from pymc_marketing.model_config import ModelConfig
 
-__all__ = ["ShiftedBetaGeoModel", "ShiftedBetaGeoModelIndividual"]
+__all__ = ["ShiftedBetaGeoModel"]
 
 
 class ShiftedBetaGeoModel(CLVModel):
@@ -756,239 +755,3 @@ class ShiftedBetaGeoModel(CLVModel):
         return retention_elasticity.transpose(
             "chain", "draw", "customer_id", "cohort", missing_dims="ignore"
         )
-
-
-class ShiftedBetaGeoModelIndividual(CLVModel):
-    """Shifted Beta Geometric model for individual customers.
-
-    Model for customer behavior in a discrete contractual setting. It assumes that:
-
-    * At the end of each period, a customer has a probability ``theta`` of renewing the contract
-      and ``1 - theta`` of cancelling
-    * The probability ``theta`` does not change over time for a given customer
-    * The probability ``theta`` varies across customers according to a Beta prior distribution
-      with hyperparameters ``alpha`` and ``beta``.
-
-    Based on [1]_.
-
-    Parameters
-    ----------
-    data: pd.DataFrame
-        DataFrame containing the following columns:
-
-        * ``customer_id``: Customer labels. There should be one unique label for each customer.
-        * ``t_churn``: Time at which the customer cancelled the contract (starting at 0).
-          It should equal ``T`` for users that have not cancelled by the end of the
-          observation period.
-        * ``T``: Maximum observed time period (starting at 0).
-    model_config: dict, optional
-        Dictionary of model prior parameters. If not provided, the model will use default priors specified in the
-        `default_model_config` class attribute.
-    sampler_config: dict, optional
-        Dictionary of sampler parameters. Defaults to None.
-
-
-    Examples
-    --------
-        .. code-block:: python
-
-            import pandas as pd
-            import pymc as pm
-
-            from pymc_extras.prior import Prior
-            from pymc_marketing.clv import ShiftedBetaGeoModelIndividual
-
-
-            data = pd.DataFrame({
-                    customer_id=[0, 1, 2, 3, ...],
-                    t_churn=[1, 2, 8, 4, 8 ...],
-                    T=[8 for x in range(len(customer_id))],
-                })
-
-            model = ShiftedBetaGeoModelIndividual(
-                model_config={
-                    "alpha": Prior("HalfNormal", sigma=10),
-                    "beta": Prior("HalfStudentT", nu=4, sigma=10),
-                },
-                sampler_config={
-                    "draws": 1000,
-                    "tune": 1000,
-                    "chains": 2,
-                    "cores": 2,
-                    "nuts_kwargs": {"target_accept": 0.95},
-                },
-            )
-
-            model.fit(data=data)
-            print(model.fit_summary())
-
-            # Predict how many periods in the future are existing customers
-            likely to cancel (ignoring that some may already have cancelled)
-            expected_churn_time = model.distribution_customer_churn_time(
-                customer_id=[0, 1, 2, 3, ...],
-            )
-            print(expected_churn_time.mean("customer_id"))
-
-            # Predict churn time for 10 new customers, conditioned on data
-            new_customers_churn_time = model.distribution_new_customer_churn_time(n=10)
-            print(new_customers_churn_time.mean("new_customer_id"))
-
-
-    References
-    ----------
[…]
