### 2026-07-14 57c76ea
remove sbg-individual and notebook
