### 2026-07-14 fff9ee4
WIP prior API deprecation and CLV regression testing
