+++ tests/test_model_config.py
@@ -12,139 +12,56 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 
-import warnings
-
 import numpy as np
 import pytest
-from pymc_extras.deserialize import (
-    DESERIALIZERS,
-    register_deserialization,
-)
 from pymc_extras.prior import Prior
 
 from pymc_marketing.hsgp_kwargs import HSGPKwargs
-from pymc_marketing.model_config import ModelConfigError, parse_model_config
+from pymc_marketing.model_config import parse_model_config
 
 
 @pytest.fixture
 def model_config():
     return {
-        # Non-nested distribution
-        "beta": {
-            "dist": "Normal",
-            "kwargs": {
-                "mu": 0.0,
-                "sigma": 1.0,
-            },
-        },
-        # Nested distribution
-        "alpha": {
-            "dist": "Normal",
-            "kwargs": {
-                "mu": {
-                    "dist": "Normal",
-                    "kwargs": {
-                        "mu": 0.0,
-                        "sigma": 1.0,
-                    },
-                },
-                "sigma": {
-                    "dist": "HalfNormal",
-                    "kwargs": {
-                        "sigma": 1.0,
-                    },
-                },
-            },
-            "dims": "channel",
-        },
-        # 2D nested distribution
-        "gamma": {
-            "dist": "Normal",
-            "kwargs": {
-                "mu": {
-                    "dist": "Normal",
-                    "kwargs": {
-                        "mu": 0.0,
-                        "sigma": 1.0,
-                    },
-                    "dims": "channel",
-                },
-                "sigma": {
-                    "dist": "HalfNormal",
-                    "kwargs": {
-                        "sigma": 1.0,
-                    },
-                    "dims": "geo",
-                },
-            },
-            "dims": ("channel", "geo"),
-        },
-        # 2D explicit kwargs
-        "delta": {
-            "dist": "Normal",
-            "kwargs": {
-                "mu": np.array([1.0]),
-                "sigma": np.array([1.0, 2.0, 3.0])[:, None],
-            },
-            "dims": ("channel", "control"),
-        },
-        # Hierarchical centered distribution
-        "hierarchical_centered": {
-            "dist": "Normal",
-            "kwargs": {
-                "mu": {
-                    "dist": "Normal",
-                    "kwargs": {
-                        "mu": 0.0,
-                        "sigma": 1.0,
-                    },
-                    "dims": "channel",
-                },
-                "sigma": {
-                    "dist": "HalfNormal",
-                    "kwargs": {
-                        "sigma": 1.0,
-                    },
-                    "dims": "geo",
-                },
-            },
-            "dims": ("channel", "geo"),
-            "centered": True,
-        },
-        # Hierarchical non-centered distribution
-        "hierarchical_non_centered": {
-            "dist": "Normal",
-            "kwargs": {
-                "mu": {"dist": "HalfNormal", "kwargs": {"sigma": 2}},
-                "sigma": {"dist": "HalfNormal", "kwargs": {"sigma": 1}},
-            },
-            "dims": "channel",
-            "centered": False,
-        },
-        # 2D Hierarchical non-centered distribution
-        "hierarchical_non_centered_2d": {
-            "dist": "Normal",
-            "kwargs": {
-                "mu": {
-                    "dist": "Normal",
-                    "kwargs": {
[…]
