+++ pymc_marketing/clv/models/gamma_gamma.py
@@ -267,9 +267,9 @@ class GammaGammaModel(BaseGammaGammaModel):
         model = GammaGammaModel(
             ),
             model_config={
-                "p": {"dist": "HalfNormal", kwargs: {}},
-                "q": {"dist": "HalfStudentT", kwargs: {"nu": 4, "sigma": 10}},
-                "v": {"dist": "HalfCauchy", kwargs: {"beta": 1}},
+                "p": Prior("HalfNormal"),
+                "q": Prior("HalfStudentT", nu=4, sigma=10),
+                "v": Prior("HalfCauchy", beta=1),
             },
             sampler_config={
                 "draws": 1000,
