+++ tests/clv/models/test_shifted_beta_geo.py
@@ -22,8 +22,6 @@ from pytensor.tensor.elemwise import Elemwise
 from pymc_marketing.clv import ShiftedBetaGeoModel
 from tests.clv.conftest import mock_sample
 
-pytestmark = pytest.mark.filterwarnings("error::FutureWarning")
-
 
 class TestShiftedBetaGeoModel:
     @classmethod
