### 2026-07-14 eb3aa6e
old model_config API and prior class removal
