+++ pymc_marketing/mmm/builders/factories.py
@@ -15,15 +15,69 @@
 
 from __future__ import annotations
 
+import copy
 import importlib
 import warnings
 from collections.abc import Mapping, MutableMapping, Sequence
 from typing import Any
 
-from pymc_extras.deserialize import deserialize
+from pymc_extras.deserialize import deserialize, register_deserialization
+from pymc_extras.prior import Prior
 
-# In order to register custom deserializers
-import pymc_marketing.prior  # noqa: F401
+
+def is_alternative_prior(data: Any) -> bool:
+    """Check if the data is a dictionary representing a Prior (alternative check)."""
+    return isinstance(data, dict) and "distribution" in data
+
+
+def deserialize_alternative_prior(data: dict[str, Any]) -> Prior:
+    """Alternative deserializer that recursively handles all nested parameters.
+
+    This handles the flat ``{"distribution": ...}`` prior format used by the MMM
+    YAML schema, where any parameter might itself be a nested prior, and also
+    extracts the ``centered`` and ``transform`` parameters.
+
+    Examples
+    --------
+    This handles cases like:
+
+    .. code-block:: yaml
+
+        distribution: Gamma
+        alpha: 1
+        beta:
+            distribution: HalfNormal
+            sigma: 1
+            dims: channel
+        dims: [brand, channel]
+
+    """
+    data = copy.deepcopy(data)
+
+    distribution = data.pop("distribution")
+    dims = data.pop("dims", None)
+    centered = data.pop("centered", True)
+    transform = data.pop("transform", None)
+    parameters = data
+
+    # Recursively deserialize any nested parameters
+    parameters = {
+        key: value if not isinstance(value, dict) else deserialize(value)
+        for key, value in parameters.items()
+    }
+
+    return Prior(
+        distribution,
+        transform=transform,
+        centered=centered,
+        dims=dims,
+        **parameters,
+    )
+
+
+# Register the alternative prior deserializer for the YAML ``{"distribution": ...}``
+# format, which pymc-extras' built-in deserializer does not handle.
+register_deserialization(is_alternative_prior, deserialize_alternative_prior)
 
 # Optional short-name registry -------------------------------------------------
 REGISTRY: dict[str, Any] = {
@@ -95,7 +149,7 @@ def build(spec: Mapping[str, Any]) -> Any:
     raw_args: Sequence[Any] = raw_kwargs.pop("args", spec.get("args", ()))
 
     # Handle specific kwargs that should be processed differently
-    special_processing_keys = ["priors", "prior"]
+    special_processing_keys = ["priors", "prior", "model_config"]
 
     # Convert list dimensions to tuples for model or effect classes
     if "dims" in raw_kwargs and isinstance(raw_kwargs["dims"], list):
@@ -122,6 +176,12 @@ def build(spec: Mapping[str, Any]) -> Any:
                     kwargs[k] = priors_dict
                 elif k == "prior" and "distribution" in v:
                     kwargs[k] = deserialize(v)
+                elif k == "model_config":
+                    # Each entry may be a prior spec ({"distribution": ...}),
+                    # an HSGP/other config dict, or a plain value. ``resolve``
+                    # deserializes prior specs and leaves everything else as-is,
+                    # since ``parse_model_config`` no longer converts dicts.
+                    kwargs[k] = {mk: resolve(mv) for mk, mv in v.items()}
                 else:
                     kwargs[k] = resolve(v)
             else:
@@ -144,7 +204,9 @@ def resolve(value):
     if isinstance(value, Mapping):
         if "class" in value:
             return build(value)
-        if "distribution" in value or "special_prior" in value:
+        # "dist" is the pymc-extras serialization key; "distribution" is the
[…]
