### 2026-07-14 1076b9a
remove `ShiftedBetaGeoModelIndividual` (#2734)
* remove sbg-individual and notebook

* remove obsolete cv dev nbs

* update dev notebooks for v1.0.0

* future warning emit
