### 2026-07-15 c04aebb
[Notebook Re-Run] mmm_sensitivity_analysis (migrate to v1.0.0 plot namespace) (#2661)
* docs(mmm): re-run & migrate mmm_sensitivity_analysis to the v1.0.0 plot namespace

Part of #2614. Migrates docs/source/notebooks/mmm/mmm_sensitivity_analysis.ipynb
to pymc>=6 / arviz>=1 and the new namespace-based MMMPlotSuite v2.

- Opt in via `mmm.plot_suite = "new"`; import `from pymc_marketing.mmm import MMM`
- `mmm.plot.sensitivity_analysis/uplift_curve/marginal_curve` ->
  `mmm.plot.sensitivity.analysis/uplift/marginal`
- New API has no `xlabel`/`ylabel`/`subplot_kwargs`: capture `(fig, axes)` and set
[…]
