### 2026-07-17 89cd381
clv ignore all warnings, address prior dep comments
