+++ tests/clv/models/test_gamma_gamma.py
@@ -23,7 +23,11 @@ from pymc_marketing.clv.models.gamma_gamma import (
     GammaGammaModel,
     GammaGammaModelIndividual,
 )
-from tests.clv.conftest import mock_fit_MAP, set_model_fit
+from tests.clv.conftest import (
+    mock_fit_MAP,
+    sample_prior_predictive_ignoring_potentials,
+    set_model_fit,
+)
 
 
 class BaseTestGammaGammaModel:
@@ -249,7 +253,7 @@ class TestGammaGammaModel(BaseTestGammaGammaModel):
             model_config=custom_model_config,
         )
         model.build_model(data=self.data)
-        fake_fit = pm.sample_prior_predictive(
+        fake_fit = sample_prior_predictive_ignoring_potentials(
             draws=1000, model=model.model, random_seed=self.rng
         )
         set_model_fit(model, fake_fit.prior)
@@ -290,6 +294,10 @@ class TestGammaGammaModel(BaseTestGammaGammaModel):
             "\nlikelihood~Potential(f(q,p,v))"
         )
 
+    # mock_fit_MAP samples the prior predictive of the Potential-based likelihood
+    @pytest.mark.filterwarnings(
+        "ignore:The effect of Potentials on other parameters is ignored"
+    )
     def test_save_load(self, mocker, tmp_path):
         model = GammaGammaModel()
         save_path = tmp_path / "test_model"
