### 2026-07-17 8a03879
fix(model_builder): update model data on repeated fit
fit() and approximate_fit() skipped build_model whenever a model already
existed, so a second fit, or a fit after an explicit build_model, silently
sampled from the data baked in at build time and ignored the new X, y.

Refresh the existing model's data via _data_setter before sampling so
repeated fits use the provided X, y. Add regression tests covering a
second fit, a fit after build_model, and approximate_fit.

Fixes #2605
