### 2026-07-18 01cbf9c
fix(mmm): sample constant tensor priors via DiracDelta (refs #1749)
Transformation.sample_prior raised when a parameter was supplied as a
constant tensor instead of a Prior, because pm.sample_prior_predictive
had no free random variable to sample.

Wrap constant params in pm.DiracDelta so PyMC produces a valid prior
dataset (with correct coords/dims) for every draw, covering the
all-constant and mixed constant/Prior cases. Distinguishing constants
from distributions mirrors _create_distributions' own dispatch (a
constant is returned by identity), so no behaviour changes when all
priors are dis
[…]
