+++ docs/source/gallery/gallery.md
@@ -336,6 +336,17 @@ Welcome to the PyMC-Marketing example gallery! This gallery provides visual navi
 :::
 ::::
 
+## Predicted Incrementality by Experimentation (PIE)
+
+::::{grid} 1 2 3 3
+:gutter: 3
+
+:::{grid-item-card} Predicting Campaign Incrementality with PIE
+:img-top: ../gallery/images/pie_example.png
+:link: ../notebooks/pie/pie_example.html
+:::
+::::
+
 ## General Tutorials
 
 ::::{grid} 1 2 3 3
