### 2026-07-18 pymc-labs/pymc-marketing#2744 @markean
[JOSS] rfm_summary gives different results for string dates with datetime_format vs pre-converted datetimes
Hi! I'm one of the reviewers of JOSS submission (https://github.com/openjournals/joss-reviews/issues/10805) and was trying the CLV workflow on v0.19.4. I passed the transaction dates as strings with `datetime_format`, and got a different summary than when I convert the column to datetime myself:

```python
import pandas as pd
from pymc_marketing.clv.utils import rfm_summary

df = pd.DataFrame({
    "id": [1, 1, 1],
    "date": ["11/01/2025", "12/20/2025", "01/15/2026"],
    "spend": [50.0, 60.0, 70.0],
})
rfm_summary(df, "id", "date", "spend", datetime_format="%m/%d/%Y")

df2 = df.assign(date=pd.to_datetime(df["date"], format="%m/%d/%Y"))
rfm_summary(df2, "id", "date", "spend")
```

The first call gives:
```
   customer_id  frequency  recency     T  monetary_value
0            1        1.0     49.0  49.0            60.0
```
while the second gives:
```
   customer_id  frequency  recency     T  monetary_value
0            1        2.0     75.0  75.0            65.0
```
It looks like the January 2026 record isn't counted in the first case. Am I supposed to convert the date to datetime before calling `rfm_summary`? If so, what is `datetime_format` meant for?

Possibly related: `rfm_train_test_split` on the same string data raises `ValueError: No data available...` though there is a transaction after my `train_period_end`.
