+++ pymc_marketing/clv/utils.py
@@ -202,24 +202,28 @@ def _find_first_transactions(
     """
     select_columns = [customer_id_col, datetime_col]
 
+    if monetary_value_col:
+        select_columns.append(monetary_value_col)
+
+    transactions = transactions[select_columns].copy()
+    transactions[datetime_col] = pandas.to_datetime(
+        transactions[datetime_col], format=datetime_format
+    )
+
     if observation_period_end is None:
         observation_period_end = transactions[datetime_col].max()
 
     if isinstance(observation_period_end, pandas.Period):
         observation_period_end = observation_period_end.to_timestamp()
     if isinstance(observation_period_end, str):
-        observation_period_end = pandas.to_datetime(observation_period_end)
-
-    if monetary_value_col:
-        select_columns.append(monetary_value_col)
+        observation_period_end = pandas.to_datetime(
+            observation_period_end, format=datetime_format
+        )
 
     if sort_transactions:
-        transactions = transactions[select_columns].sort_values(select_columns).copy()
+        transactions = transactions.sort_values(select_columns)
 
     # convert date column into a DateTimeIndex for time-wise grouping and truncating
-    transactions[datetime_col] = pandas.to_datetime(
-        transactions[datetime_col], format=datetime_format
-    )
     transactions = (
         transactions.set_index(datetime_col).to_period(time_unit).to_timestamp()
     )
@@ -334,7 +338,8 @@ def rfm_summary(
     """
     if observation_period_end is None:
         observation_period_end_ts = (
-            pandas.to_datetime(transactions[datetime_col].max(), format=datetime_format)
+            pandas.to_datetime(transactions[datetime_col], format=datetime_format)
+            .max()
             .to_period(time_unit)
             .to_timestamp()
         )
@@ -487,9 +492,6 @@ def rfm_train_test_split(
         and *monetary_value* if specified
 
     """
-    if test_period_end is None:
-        test_period_end = transactions[datetime_col].max()
-
     transaction_cols = [customer_id_col, datetime_col]
     if monetary_value_col:
         transaction_cols.append(monetary_value_col)
@@ -498,6 +500,9 @@ def rfm_train_test_split(
     transactions[datetime_col] = pandas.to_datetime(
         transactions[datetime_col], format=datetime_format
     )
+    if test_period_end is None:
+        test_period_end = transactions[datetime_col].max()
+
     test_period_end = pandas.to_datetime(test_period_end, format=datetime_format)
     train_period_end = pandas.to_datetime(train_period_end, format=datetime_format)
 
