### 2026-07-20 pymc-labs/pymc-marketing#2744 @juanitorduz
Thank you @markean 🙏 . We merged a fix (https://github.com/pymc-labs/pymc-marketing/pull/2745) into our `v1.0.0` branch as we are preparing for a stable release soon.
