### 2026-07-21 656900d
docs: address JOSS review feedback on documentation
