+++ docs/source/guide/benefits/why_bayesian.md
@@ -4,11 +4,11 @@ Data such as customer transactions and advertising expenditure can be low resolu
 
 Bayesian approaches allow your team’s valuable domain expertise to be incorporated in the modelling process in the form of Bayesian prior beliefs. This can make a massive difference - especially when data are noisy and uncertain. Rather than your model contorting itself in strange ways to best fit complex and noisy data, Bayesian priors can keep your model right, leading to much more sensible insights that fit with your domain expertise.
 
-Frequentist approaches can require two or more years worth of historical data. Bayesian methods can work with very short-run data, meaning you don't have to wait a long time collecting data before getting insights. The certainty in your estimates will grow as your dataset increases.
+Frequentist approaches can require substantial historical data. Bayesian methods can work with very short-run data, meaning you don't have to wait a long time collecting data before getting insights. The certainty in your estimates will grow as your dataset increases.
 
 Bayesian approaches excel at modelling hierarchical or nested data. This is particularly useful if you have just launched a new product, or operate in a new region, or are dealing with a new cohort or demographic of customers and don’t have many observations. Bayesian hierarchical models allow information learnt about some categories to intelligently inform you about novel categories.
 
-Bayesian methods provide *the* way to make decisions under uncertainty. This enables you to generate and predict future scenarios - knowing how certain or not you are of those predictions. You can run optimisation processes to decide what actions to take in the future, fully taking uncertainty into account. This helps manage your risk.
+Bayesian methods provide *a principled* way to make decisions under uncertainty. This enables you to generate and predict future scenarios - knowing how certain or not you are of those predictions. You can run optimisation processes to decide what actions to take in the future, fully taking uncertainty into account. This helps manage your risk.
 
 ## Find out more
 To find out more, click through the video thumbnail below to watch the talk [Solving Real-World Business Problems with Bayesian Modeling](https://www.youtube.com/watch?v=twpZhNqVExc) by Thomas Wiecki, recorded at PyData London 2022.
