### 2026-07-21 7f84894
docs: clarify future_t parameter is always in months
