### 2026-07-21 7689b40
fix(clv): correct daily lifetime value scaling
Use the average Gregorian month length when converting monthly CLV horizons to days. Add a regression that verifies a 12-month horizon requests 365.25 days.
