### 2026-07-21 59a8320
Add PIE notebook (#2743)
* Add PIE notebook

* Fix CI/CD for variable importance

* Migrate PIE notebook to pymc 6 / arviz 1; exclude from mock notebook CI

* Respond to PR comments

---------

Co-authored-by: Juan Orduz <juanitorduz@gmail.com>
