### 2026-07-21 83c7a42
address JOSS review feedback on paper
