+++ paper/paper.bib
@@ -9,8 +9,8 @@
 @online{pymclabs2023bolt,
   author = {{PyMC Labs}},
   title = {Bolt's Success Story with PyMC-Marketing},
-  url = {https://www.pymc-marketing.io/en/stable/},
-  urldate = {2026-05-14},
+  url = {https://bolt.eu/en/blog/budgeting-with-bayesian-models-pymc-marketing/},
+  urldate = {2026-07-20},
   year = {2023}
 }
 
@@ -126,7 +126,7 @@
 }
 
 @misc{facebook2022robyn,
-  title={Robyn: Marketing Mix Modeling},
+  title={{Robyn: Marketing Mix Modeling}},
   author={{Meta Platforms Inc.}},
   year={2022},
   version={3.10.0},
@@ -134,7 +134,7 @@
 }
 
 @misc{google2023meridian,
-  title={Meridian: Bayesian Media Mix Modeling at Scale},
+  title={{Meridian: Bayesian Media Mix Modeling at Scale}},
   author={{Google Inc.}},
   year={2023},
   version={0.1.0},
@@ -177,7 +177,7 @@
 @software{Seyboldt_nutpie,
   author = {Seyboldt, Adrian and {PyMC Developers}},
   license = {MIT},
-  title = {{nutpie}},
+  title = {{Nutpie}},
   year = {2023},
   version = {0.5.0},
   url = {https://github.com/pymc-devs/nutpie}
@@ -207,7 +207,7 @@
 }
 
 @article{sklearn2011,
-  title={Scikit-learn: Machine learning in Python},
+  title={{Scikit-learn: Machine learning in Python}},
   author={Pedregosa, Fabian and Varoquaux, Gaël and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and others},
   journal={the Journal of machine Learning research},
   volume={12},
