+++ paper/paper.md
@@ -66,19 +66,19 @@ bibliography: paper.bib
 
 # Summary
 
-PyMC-Marketing is a comprehensive Python library implementing Bayesian marketing analytics, built on PyMC [@pymc2023]. Commercial marketing analytics tools typically provide limited transparency into their models, while open-source alternatives like Meta's Robyn and Google's Meridian focus primarily on media mix modeling [@facebook2022robyn; @google2023meridian]. PyMC-Marketing provides a unified framework spanning multiple marketing domains, including: Marketing Mix Models (MMMs), Customer Lifetime Value (CLV) analysis, Bass Diffusion Models, and Customer Choice Models. All outputs include full posterior distributions rather than point estimates, enabling explicit risk assessment in business decisions. Key capabilities include time-varying media coefficients via Hilbert space Gaussian process approximations [@solin2020hilbert], experimental calibration integrating lift tests into model likelihood, hierarchical BTYD models [@fader2020customer] for CLV, and budget optimization with business constraints. The library is available on PyPI and conda-forge, with comprehensive documentation and example notebooks at https://www.pymc-marketing.io.
+PyMC-Marketing is a comprehensive Python library implementing Bayesian marketing analytics, built on PyMC [@pymc2023]. Commercial marketing analytics tools typically provide limited transparency into their models, while open-source alternatives like Meta's Robyn and Google's Meridian focus primarily on media mix modeling [@facebook2022robyn; @google2023meridian]. PyMC-Marketing provides a unified framework spanning multiple marketing domains, including: Marketing Mix Models (MMMs), Customer Lifetime Value (CLV) analysis, Bass Diffusion Models, and Customer Choice Models. All outputs include full posterior distributions rather than point estimates, enabling explicit risk assessment in business decisions. Key capabilities include time-varying media coefficients via Hilbert space Gaussian process approximations [@solin2020hilbert], experimental calibration integrating lift tests into model likelihood, hierarchical Buy-Till-You-Die (BTYD) models [@fader2020customer] for CLV, and budget optimization with business constraints. The library is available on PyPI and conda-forge, with comprehensive documentation and example notebooks at https://www.pymc-marketing.io.
 
 # Statement of Need
 
 Marketing organizations struggle to attribute sales outcomes to specific marketing activities across multiple touchpoints and delayed conversion effects. Existing solutions suffer from: (1) black-box proprietary models with limited customization; (2) oversimplified approaches failing to capture marketing dynamics; and (3) lack of uncertainty quantification for high-stakes decisions. The target audience includes marketing data scientists and analysts in industry, academic researchers in causal inference and marketing science, and practitioners building production marketing measurement systems.
 
-PyMC-Marketing addresses these gaps by bridging marketing science research and practical applications. It operationalizes advanced Bayesian methods—hierarchical modeling, experimental calibration, and uncertainty quantification—within a user-friendly, scikit-learn compatible API. Key innovations include time-varying coefficients using modern Gaussian process approximations optimized for marketing contexts, and a novel experimental calibration framework that integrates lift test results directly into model likelihood. While frequentist approaches like Robyn provide bootstrap-based intervals, all PyMC-Marketing outputs include full Bayesian posterior distributions, enabling decision-makers to assess risk explicitly.
+PyMC-Marketing addresses these gaps by bridging marketing science research and practical applications. It operationalizes Bayesian methods—hierarchical modeling, experimental calibration, and uncertainty quantification—within a user-friendly, scikit-learn-style API. Key innovations include time-varying coefficients using modern Gaussian process approximations optimized for marketing contexts, and a novel experimental calibration framework that integrates lift test results directly into model likelihood.
 
 # State of the Field
 
[…]
