### 2026-07-27 0a25142
feat(mmm): add DataVarMuEffect, MediaMuEffect, ControlMuEffect for xarray data referencing
Three new MuEffect subclasses enabling media and control effects that
read data directly from named variables in the training xr.Dataset.

- DataVarMuEffect: abstract base with create_data/set_data defaults
- MediaMuEffect: applies MediaTransformation to a named variable,
  aggregates over configurable channel_dim
- ControlMuEffect: configurable prior per control variable, auto-sums
  extra dimensions

Also extends the Model protocol with xarray_dataset property so all
MuEffects can access the canonical 
[…]
