+++ pymc_marketing/mmm/additive_effect.py
@@ -104,6 +104,107 @@ Tips for custom components
     with dims ``("date", *mmm.dims)``. If you need broadcasting, use
     ``pymc_extras.prior.create_dim_handler`` as shown above.
   - In ``set_data``, update the data variables when dates/dims change.
+
+Built-in data-referencing effects
+----------------------------------
+
+The module provides ready-to-use ``MuEffect`` subclasses that read data
+directly from the training ``xr.Dataset``.
+
+``DataVarMuEffect``
+    Abstract base for effects that reference named variables in the Dataset.
+    Subclasses implement ``create_effect``; ``create_data`` and ``set_data``
+    are provided.
+
+``MediaMuEffect(DataVarMuEffect)``
+    Applies a ``MediaTransformation`` (adstock + saturation) to a named
+    media variable, then aggregates over ``channel_dim``.
+
+``ControlMuEffect(DataVarMuEffect)``
+    Applies a configurable prior coefficient to each control variable,
+    automatically summing extra dimensions.
+
+Example: multi-granularity media and controls
+..............................................
+
+.. code-block:: python
+
+    from pymc_marketing.mmm import MMM, GeometricAdstock, LogisticSaturation
+    from pymc_marketing.mmm.additive_effect import (
+        MediaMuEffect,
+        ControlMuEffect,
+    )
+    from pymc_marketing.mmm.media_transformation import MediaTransformation
+
+    # X is an xr.Dataset with:
+    #   media_product:      (date, product, product-channel)
+    #   media_geo:          (date, geo, geo-channel)
+    #   control_national:   (date,)
+    #   control_product:    (date, product)
+
+    mmm = (
+        MMM(
+            date_column="date",
+            channel_columns=["tv", "digital"],
+            dims=("product", "geo"),
+            adstock=GeometricAdstock(l_max=8),
+            saturation=LogisticSaturation(),
+        )
+        .add_mu_effect(
+            MediaMuEffect(
+                data_vars=["media_product"],
+                media_transformation=MediaTransformation(
+                    adstock=GeometricAdstock(l_max=8),
+                    saturation=LogisticSaturation(),
+                    adstock_first=True,
+                    dims=("product", "product-channel"),
+                ),
+                channel_dim="product-channel",
+                prefix="product_media",
+            )
+        )
+        .add_mu_effect(
+            MediaMuEffect(
+                data_vars=["media_geo"],
+                media_transformation=MediaTransformation(
+                    adstock=GeometricAdstock(l_max=8),
+                    saturation=LogisticSaturation(),
+                    adstock_first=True,
+                    dims=("geo", "geo-channel"),
+                ),
+                channel_dim="geo-channel",
+                prefix="geo_media",
+            )
+        )
+        .add_mu_effect(
+            ControlMuEffect(
+                data_vars=["control_national"],
+                prefix="national_ctrl",
+            )
+        )
+        .add_mu_effect(
+            ControlMuEffect(
+                data_vars=["control_product"],
+                prefix="product_ctrl",
+            )
+        )
+    )
+
+    mmm.fit(X, y)
+
+Each grain uses a distinct dimension name (``"product-channel"`` vs
+``"geo-channel"``) to avoid xarray's coordinate union and the ``NaN``
+values it would produce.  ``ControlMuEffect`` uses a scalar
+``Prior("Normal", ...)`` by default, broadcasting across all dimensions;
+pass ``Prior("Normal", mu=0, sigma=2, dims="product")`` for per-product
+coefficients.
+
+.. note::
+
+    ``MediaMuEffect`` does not apply any automatic scaling.  Media data
+    should be pre-scaled (e.g. max-scaling) before being placed in the
[…]
