### 2026-07-27 307a611
fix review feedback: set_data on base, Field(min_length=1), rm dead effect_dims
- Lift set_data from MediaMuEffect/ControlMuEffect to DataVarMuEffect
  base to eliminate byte-for-byte duplicate
- Add Annotated[Field(min_length=1)] to data_vars so empty lists
  are caught at construction
- Remove stale effect_dims key from MediaMuEffect.to_dict since
  it is derived from media_transformation.dims at runtime
- Update DataVarMuEffect docstring to reflect set_data is concrete
