### 2026-07-27 afeed1a
Merge branch 'main' into v1.0.0 - resolve merge conflicts
