+++ .github/workflows/bot-prs.yml
@@ -11,7 +11,7 @@ jobs:
   label-no-releasenotes:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v6
+      - uses: actions/checkout@v7
       - name: Label no releasenotes
         run: |
           number=${{ github.event.pull_request.number }}
