### 2026-07-27 481c114
Fix DataTree and arviz API compatibility for customer_choice modules
