### 2026-07-27 0807d41
Fix DataTree.extend -> update for BayesianBLP compatibility
