### 2026-07-27 4263569
Merge resolved conflicts into v1.0.0 (#2759)
* Merge branch 'main' into v1.0.0 - resolve merge conflicts

* Fix DataTree and arviz API compatibility for customer_choice modules

* Fix DataTree.extend -> update for BayesianBLP compatibility
