### 2026-07-28 6283f79
fix(plot): use matplotlib date converter for datetime x-axis (#2685)
* fix(plot): use matplotlib date converter for datetime x-axis

plot_curve drew sample lines with pandas period ordinals while the HDI
band used matplotlib date ordinals, producing broken x-axis tick labels
on datetime coordinates. Pass x_compat=True so pandas uses the
matplotlib date converter. No-op for numeric x-axes.

Fixes #2682

* docs(mmm): refresh mmm_components event plot with corrected date axis

Re-run cell 68 so the AsymmetricGaussianBasis comparison shows proper
date ticks, and drop the fig.autofmt_xdate(rotation
[…]
