### 2026-07-28 f2ecf67
Final merge — resolve v1.0.0 conflicts with main (#2764)
* Add v1.0.0 branch to CI triggers (#2610)

* Add v1.0.0 branch to CI triggers (#2618)

* Add v1.0.0 to codecov branch tracking

* chore: Update UML Diagrams (#2608)

Co-authored-by: williambdean <57733339+williambdean@users.noreply.github.com>

* chore(triage): add no releasenotes label to triage agent (#2618)

The triage agent now recognizes the 'no releasenotes' label for issues
about internal/process-only changes (CI/CD, test infrastructure,
automation, version bumps, agent config) that should be excluded from
release notes.

Key gua
[…]
