+++ .github/workflows/docs.yml
@@ -45,7 +45,7 @@ jobs:
     steps:
       - uses: actions/checkout@v7
       - name: Install uv
-        uses: astral-sh/setup-uv@v7
+        uses: astral-sh/setup-uv@v8.3.2
       - name: Set up Python
         run: uv python install 3.12
       - name: Set up environment
