### 2026-07-28 5cb05ef
feat(mmm): DataVarMuEffect, MediaMuEffect, ControlMuEffect for xarray data referencing (#2757)
* feat(mmm): add DataVarMuEffect, MediaMuEffect, ControlMuEffect for xarray data referencing

Three new MuEffect subclasses enabling media and control effects that
read data directly from named variables in the training xr.Dataset.

- DataVarMuEffect: abstract base with create_data/set_data defaults
- MediaMuEffect: applies MediaTransformation to a named variable,
  aggregates over configurable channel_dim
- ControlMuEffect: configurable prior per control variable, auto-sums
  extra dimensions

Also 
[…]
