+++ pymc_marketing/mmm/__init__.py
@@ -14,6 +14,11 @@
 """Marketing Mix Models (MMM)."""
 
 from pymc_marketing.mmm import preprocessing, validating
+from pymc_marketing.mmm.additive_effect import (
+    ControlMuEffect,
+    DataVarMuEffect,
+    MediaMuEffect,
+)
 from pymc_marketing.mmm.components.adstock import (
     AdstockTransformation,
     BinomialAdstock,
@@ -85,8 +90,10 @@ __all__ = [
     "AdstockTransformation",
     "BinomialAdstock",
     "BudgetOptimizerWrapper",
+    "ControlMuEffect",
     "CovFunc",
     "DataDerivedScaling",
+    "DataVarMuEffect",
     "DelayedAdstock",
     "FancyLinearRegression",
     "FixedScaling",
@@ -102,6 +109,7 @@ __all__ = [
     "MMMPlotSuiteFacade",
     "MediaConfig",
     "MediaConfigList",
+    "MediaMuEffect",
     "MediaTransformation",
     "MichaelisMentenSaturation",
     "MonthlyFourier",
