### 2026-07-28 d710dc0
v1.0.0 (#2677)
* Add v1.0.0 branch to CI triggers (#2618)

* Bump version to 1.0.0.dev0

* Migrate from conda to uv for development environment (#2611)

* Remove deprecated serialization shims kept for Legacy MMM (#2599)

* rm

* rm test

* Remove TestDeprecatedShimsRemoved and importlib from test_saturation

* Fix broken saturation_from_dict import in test_link.py

---------

Co-authored-by: Will Dean <wd60622@gmail.com>

* fix: pin pymc-extras<0.11.0 and matplotlib<3.11 on v1.0.0 branch (#2629)

* fix: pin pymc-extras<0.11.0 on v1.0.0 branch

Backports the pin from #2623 (main). pymc-extras 
[…]
