+++ tests/clv/models/test_basic.py
@@ -15,17 +15,9 @@ import numpy as np
 import pandas as pd
 import pymc as pm
 import pytest
-from arviz import InferenceData, from_dict
+import xarray as xr
 from pymc_extras.prior import Prior
 
-from pymc_marketing.clv.models import (
-    BetaGeoBetaBinomModel,
-    BetaGeoModel,
-    GammaGammaModel,
-    ModifiedBetaGeoModel,
-    ParetoNBDModel,
-    ShiftedBetaGeoModel,
-)
 from pymc_marketing.clv.models.basic import CLVModel
 from pymc_marketing.model_builder import DifferentModelError
 from tests.clv.conftest import mock_fit_MAP, mock_sample, set_model_fit
@@ -36,7 +28,7 @@ class CLVModelTest(CLVModel):
 
     def __init__(
         self,
-        data=None,
+        data: pd.DataFrame | None = None,
         model_config=None,
         sampler_config: dict | None = None,
     ):
@@ -44,11 +36,11 @@ class CLVModelTest(CLVModel):
             data = pd.DataFrame({"y": np.random.randn(10)})
 
         super().__init__(
-            data=data,
             model_config=model_config,
             sampler_config=sampler_config,
             non_distributions=[],
         )
+        self.data = data
 
     @property
     def default_model_config(self):
@@ -90,11 +82,12 @@ class CLVModelForLoadTest(CLVModelTest):
     ):
         CLVModel.__init__(
             self,
-            data=data,
             model_config=model_config,
             sampler_config=sampler_config,
             non_distributions=[],
         )
+        if data is not None:
+            self.data = data
 
 
 @pytest.fixture(scope="module")
@@ -105,8 +98,8 @@ def posterior():
     )  # shape convention: (chain, draw, *shape)
 
     # Create a dictionary for posterior
-    posterior_dict = {"theta": posterior_samples}
-    return from_dict(posterior=posterior_dict)
+    posterior_dict = xr.Dataset({"theta": posterior_samples})
+    return xr.DataTree.from_dict({"/posterior": posterior_dict})
 
 
 class TestCLVModel:
@@ -114,7 +107,7 @@ class TestCLVModel:
         model = CLVModelTest()
         assert model.__repr__() == "CLVModelTest"
 
-        model.build_model()
+        model.build_model(model.data)
         assert model.__repr__() == "CLVModelTest\nx ~ Normal(0, 1)\ny ~ Normal(x, 1)"
 
     def test_fit_mcmc(self, mocker):
@@ -123,26 +116,32 @@ class TestCLVModel:
         mocker.patch("pymc.sample", mock_sample)
 
         idata = model.fit(
+            data=model.data,
             tune=5,
             chains=2,
             draws=10,
             compute_convergence_checks=False,
         )
-        assert isinstance(idata, InferenceData)
-        assert len(idata.posterior.chain) == 2
-        assert len(idata.posterior.draw) == 10
-        assert model.fit_result is idata.posterior
+        assert isinstance(idata, xr.DataTree)
+        assert idata["/posterior"].to_dataset().sizes["chain"] == 2
+        assert idata["/posterior"].to_dataset().sizes["draw"] == 10
+        assert model.fit_result.equals(idata["/posterior"].to_dataset())
+        assert isinstance(model.fit_result, xr.Dataset)
 
     def test_fit_map(self, mocker):
         model = CLVModelTest()
 
         mocker.patch("pymc_marketing.clv.models.basic.CLVModel._fit_MAP", mock_fit_MAP)
-        idata = model.fit(method="map")
+        idata = model.fit(
+            data=model.data,
+            method="map",
+        )
 
-        assert isinstance(idata, InferenceData)
-        assert len(idata.posterior.chain) == 1
-        assert len(idata.posterior.draw) == 1
[…]
