+++ .github/workflows/pypi.yml
@@ -35,7 +35,7 @@ jobs:
       - uses: actions/checkout@v7
         with:
           fetch-depth: 0
-      - uses: actions/setup-python@v6
+      - uses: actions/setup-python@v7
         with:
           python-version: "3.12"
       - name: Build the sdist and the wheel
@@ -91,7 +91,7 @@ jobs:
         with:
           skip-existing: true
           repository-url: https://test.pypi.org/legacy/
-      - uses: actions/setup-python@v6
+      - uses: actions/setup-python@v7
         with:
           python-version: "3.12"
       - name: Test pip install from test.pypi
