### 2026-07-28 0679fe0
docs(mmm): add advanced geo-level funnel MMM example via DataVarMuEffect
Extends the funnel-aware MMM example to a setup closer to production work:
two upper-funnel channels feeding a shared latent demand pool across a geo
panel, two noisy demand proxies (lower-funnel spend with its loading fixed
at one, plus branded search volume with a free loading), and seasonal
confounding between category demand and the media plans.

The custom effect subclasses DataVarMuEffect (added in #2757), so it only
implements create_effect and to_dict -- create_data and set_data come for
free, and the mediator dat
[…]
