### 2026-07-28 20e4be6
docs(mmm): make the funnel DAG causal and improve figure readability
Reworks the advanced funnel example so that lower-funnel spend is a genuine
cause of the target rather than a terminal proxy, and makes every figure
readable at a glance.

Causal structure
- Adds an exogenous lower-funnel budget B. Structural lower-funnel spend is
  now M* = D + lambda_g * B, observed as `lower_spend` with reporting noise,
  and M* drives the target through its own adstock/saturation. The mediation
  chain is literally U -> D -> M* -> Y.
- `search_volume` stays a pure indicator of latent demand with no edge t
[…]
