### 2026-07-28 9f8b53b
docs(mmm): address review - Naive B+, d-separation analysis, ROAS conventions
Responds to the detailed review on PR #2769:

- Fix the MultiIndex assignment that blanked the mean |error| row (1.1)
- Correct the include_carryover note and demonstrate the flag with
  frequency="yearly", where it genuinely bites (1.3)
- Add Naive B+ (category_demand as a control), the cheapest formally valid
  total-effect estimator, fully integrated into every comparison; its
  failure (gamma_control ~ 0) becomes the notebook's sharpest lesson:
  a valid adjustment set is not a sufficient adjustment (2.1)
- Name 
[…]
