### 2026-07-29 abebb04
Funnel Effect Example via MuEffects (#2765)
* docs(mmm): add funnel-aware MMM example via a custom MuEffect

Adds mmm_funnel_mueffect.ipynb: upper- to lower-funnel mediation modelled
with a custom MuEffect and a second likelihood on the mediator, plus the
gallery entry under Causal Inference.

The DGP distinguishes latent lower-funnel demand D_t from observed spend
M_t = D_t + eps^M, with the target responding to D_t, and the in-sample R^2
rescale reads each model's own target_scale.

Co-authored-by: Cursor <cursoragent@cursor.com>

* docs(mmm): address review feedback on funnel MuEffect examp
[…]
