+++ docs/source/gallery/gallery.md
@@ -226,6 +226,11 @@ Welcome to the PyMC-Marketing example gallery! This gallery provides visual navi
 :link: ../notebooks/mmm/mmm_upper_funnel_causal_approach.html
 :::
 
+:::{grid-item-card} Funnel-Aware MMM: Upper to Lower-Funnel Mediation via a custom `MuEffect`
+:img-top: ../gallery/images/mmm_funnel_mueffect.png
+:link: ../notebooks/mmm/mmm_funnel_mueffect.html
+:::
+
 :::{grid-item-card} Causal Reasoning and Discovery
 :img-top: ../gallery/images/mmm_causal_reasoning_and_discovery.png
 :link: ../notebooks/mmm/mmm_causal_reasoning_and_discovery.html
