### 2026-07-29 86c3b27
fix(bass): derive forecast shape from dims and keep T leading
- _data_setter builds the no-observed placeholder from y_obs's declared dims instead of assuming T is axis 0
- to_bass_dataset transposes an xr.Dataset so T leads, so a (product, T) dataset fits and forecasts
- tests: transposed-dataset forecast, and assert model_config equality on the save/load round-trip
