### 2026-07-29 c803d4d
docs(mmm): make the growth trend a material confounder in the advanced funnel example
Address the third review round on the funnel MuEffect advanced notebook:
differentiate the DGP trends (TV 1.20, social 0.15, category 1.00, noise 0.30)
so the backdoor the DAG section is organised around is live and measured, add
the residual-correlation measurement cell (raw / after Fourier / after
Fourier+t), rework the gamma_control diagnostic (pointwise-average implied
slope, distance-from-implied and distance-from-zero columns), compute the
U2-treatment d-separation symmetry instead of asserting it, prin
[…]
