+++ pyproject.toml
@@ -27,6 +27,7 @@ classifiers = [
 dependencies = [
     "arviz>=1.2.0,<2.0",
     "arviz-plots[matplotlib]>=1.2.0,<2.0",
+    "impulso==0.0.8",
     "matplotlib>=3.5.1,<3.11",
     "narwhals",
     "numpy>=2.0,<3.0",
