### 2026-07-30 8d70eeb
docs(mmm): make the DAG figure show the node the analysis turns on
The graphviz `color` attribute sets the pen on an unfilled node, so `N` and
`lower_spend` were drawn white on a white canvas: the unobserved common cause
whose presence changes every d-separation verdict was invisible in the one
figure the whole section reads from. Give the drawing a single convention
instead of patching the colour -- dashed outline means no dataset carries it
(`N` and `D`), filled means observed, and each edge style means one thing --
and state it where the figure is read.

Two claims the section still asserte
[…]
