### 2026-07-30 322c22f
ci(docs): skip RTD PR preview builds when no docs-relevant files changed
Ports the post_checkout skip logic already running in pymc-labs/causalpy.
On external (PR) builds, diff against origin/main over the paths that can
affect rendered docs (docs/, .readthedocs.yml, pymc_marketing/, scripts/)
and exit 183 when nothing relevant changed, which RTD treats as a cancelled
build rather than a failure. pymc_marketing/data/ is excluded since data
files do not affect the rendered output. Pushes to main always rebuild.

Co-Authored-By: Claude <noreply@anthropic.com>
