### 2026-07-30 dfe53aa
ci(docs): mirror docs.yml path filter in RTD skip logic
Addresses review feedback on the skip rule:

- Add pyproject.toml. RTD installs the package with the [docs] extra from
  it, so dependency/metadata edits can change the Sphinx environment.
- Drop the pymc_marketing/data/ exclusion. That tree is a real Python
  package (fivetran, idata), not data fixtures, so changes to it do affect
  autodoc output. The CSV fixtures live in the top-level data/ directory,
  which is already outside the watched paths.
- Narrow scripts/ to scripts/generate_gallery.py, the only script the docs
  build actuall
[…]
