+++ .readthedocs.yml
@@ -17,25 +17,22 @@ build:
       # Skip PR preview builds when no docs-relevant files changed.
       # exit 183 cancels the build (shown as green on the PR, not a failure).
       # Pushes to main are unaffected — live docs always rebuild.
+      #
+      # The path list mirrors the `changes` filter in .github/workflows/docs.yml
+      # so the RTD preview and the GH Actions `Docs` job agree on what counts as
+      # docs-relevant. Keep the two in sync.
       # Ref: https://docs.readthedocs.com/platform/stable/guides/build/skip-build.html
       - git fetch origin main --depth=1
       - |
         if [ "$READTHEDOCS_VERSION_TYPE" = "external" ]; then
-          CHANGED_FILES=$(git diff --name-only origin/main -- docs/ .readthedocs.yml pymc_marketing/ scripts/ 2>/dev/null || true)
-          SHOULD_SKIP=1
+          CHANGED_FILES=$(git diff --name-only origin/main -- \
+            .readthedocs.yml \
+            pyproject.toml \
+            docs/ \
+            pymc_marketing/ \
+            scripts/generate_gallery.py 2>/dev/null || true)
 
-          for FILE in $CHANGED_FILES; do
-            case "$FILE" in
-              pymc_marketing/data/*)
-                ;;
-              *)
-                SHOULD_SKIP=0
-                break
-                ;;
-            esac
-          done
-
-          if [ "$SHOULD_SKIP" -eq 1 ]; then
+          if [ -z "$CHANGED_FILES" ]; then
             exit 183;
           fi
         fi
