### 2026-07-30 ff44b0e
feat(mmm): optimizable effect levers via names-based BudgetOptimizer.optimizable_vars
Fill the seam #2654 reserved for OptimizableMuEffect integration (#2621),
following that PR's philosophy -- "configurable variable and dimension names
replace hardcoded strings". The optimizer generalizes from one optimizable
variable name (channel_data_var) to N, and stays a pure pm.Model graph tool:

- BudgetOptimizer gains `optimizable_vars: {pm.Data name: native (low, high)
  bounds | None}`. Each variable's dim/size/coords are read off the model
  graph (one non-date dim required); its flat-vector tail s
[…]
