+++ pymc_marketing/mmm/additive_effect.py
@@ -106,9 +106,11 @@ Tips for custom components
   - In ``set_data``, update the data variables when dates/dims change.
 """
 
+import warnings
 from abc import ABC, abstractmethod
 from typing import Any, Protocol
 
+import numpy as np
 import numpy.typing as npt
 import pandas as pd
 import pymc as pm
@@ -116,6 +118,7 @@ import pymc.dims as pmd
 import pytensor.xtensor as ptx
 import xarray as xr
 from pydantic import Field, InstanceOf
+from pymc_extras.prior import Prior
 from pytensor.xtensor.type import XTensorVariable
 
 from pymc_marketing.mmm.events import EventEffect, days_from_reference
@@ -283,6 +286,36 @@ class MuEffect(SerializableBaseModel, ABC):
             Group name to xarray Dataset mapping.
         """
         return {}
+
+
+class OptimizableMuEffect(MuEffect, ABC):
+    """A :class:`MuEffect` whose own lever can be optimized jointly with media budgets.
+
+    The lever is the effect's ``f"{prefix}_data"`` node (a ``pm.Data`` variable
+    registered in :meth:`MuEffect.create_data` with a single non-date dim).
+    ``MMM.budget_optimizer`` translates each optimizable effect into an
+    ``optimizable_vars`` entry -- the node name plus the effect's native
+    :attr:`lever_bounds` -- and the
+    :class:`~pymc_marketing.mmm.budget_optimizer.BudgetOptimizer` co-optimizes
+    that variable purely by name on the model graph, alongside the media
+    budgets. The optimal values come back in ``result.optimized_vars``.
+
+    The lever never enters the media budget-sum constraint; its economic cost
+    (if any) belongs in the model response. Pair the optimization with
+    ``response_variable="total_response_original_scale"`` so the effect's
+    contribution enters the objective.
+    """
+
+    @property
+    def lever_bounds(self) -> list[tuple[float | None, float | None]] | None:
+        """Per-lever ``(low, high)`` bounds in the lever's native units.
+
+        ``None`` (default) leaves the lever unbounded. Effects with a naturally
+        bounded lever (e.g. a discount fraction) should override this with one
+        ``(low, high)`` tuple per lever entry, in the lever dim's coordinate
+        order.
+        """
+        return None
 
 
 class FourierEffect(MuEffect):
@@ -813,3 +846,422 @@ def _register_event_additive_effect() -> None:
 
 
 _register_event_additive_effect()
+
+
+class DiscountedEventEffect(OptimizableMuEffect):
+    r"""Promotional event effect with a log-log discount-depth lever.
+
+    Each event window (e.g. Black Friday, Summer Sale) has a known date range.
+    The optimizer decides *how deep* to discount for each event.  The lever is
+    the **discount fraction** :math:`d_k \in [0, 1]` (e.g. ``0.20`` for a 20 %
+    discount), optimized directly in native units between
+    :attr:`discount_min` and :attr:`discount_max`.
+
+    The per-event lift uses a full revenue-retention specification:
+
+    .. math::
+
+        \text{lift}_k = \beta_k \cdot \ln(1 + d_k) \cdot (1 - d_k)
+                        - d_k \cdot r_k
+
+    * :math:`\ln(1 + d_k)` — volume uplift (log-linear elasticity).
+    * :math:`(1 - d_k)` — price retention; at 100 % discount the business
+      collects nothing per unit, so the :math:`\beta` term vanishes.
+    * :math:`-d_k \cdot r_k` — margin forgone on the existing customer base:
+      a fraction :math:`d_k` of the baseline per-period revenue :math:`r_k` is
+      sacrificed.
+
[…]
