+++ tests/mmm/test_additive_effect.py
@@ -19,10 +19,14 @@ import pytest
 import xarray as xr
 from pymc_extras.prior import Prior
 
+from pymc_marketing.mmm import MMM
 from pymc_marketing.mmm.additive_effect import (
+    DiscountedEventEffect,
     FourierEffect,
     LinearTrendEffect,
 )
+from pymc_marketing.mmm.components.adstock import GeometricAdstock
+from pymc_marketing.mmm.components.saturation import LogisticSaturation
 from pymc_marketing.mmm.fourier import MonthlyFourier, WeeklyFourier, YearlyFourier
 from pymc_marketing.mmm.linear_trend import LinearTrend
 from pymc_marketing.serialization import DeserializationContext, serialization
@@ -477,3 +481,147 @@ class TestEventAdditiveEffectRoundtrips:
 )
 def test_additive_effect_type_registered(type_key):
     assert type_key in serialization._registry, f"{type_key} not registered"
+
+
+def _make_discount_events() -> pd.DataFrame:
+    return pd.DataFrame(
+        {
+            "name": ["black_friday", "summer_sale"],
+            "start_date": ["2025-03-03", "2025-01-20"],
+            "end_date": ["2025-03-24", "2025-02-10"],
+            "discount_pct": [0.30, 0.20],
+        }
+    )
+
+
+class TestDiscountedEventEffectValidation:
+    def test_missing_columns_raise(self):
+        with pytest.raises(ValueError, match="missing required columns"):
+            DiscountedEventEffect(
+                df_events=pd.DataFrame({"name": ["a"], "start_date": ["2024-01-01"]}),
+                prefix="promo",
+            )
+
+    def test_inverted_bounds_raise(self):
+        with pytest.raises(ValueError, match="discount_min"):
+            DiscountedEventEffect(
+                df_events=_make_discount_events(),
+                prefix="promo",
+                discount_min=0.5,
+                discount_max=0.2,
+            )
+
+    def test_discount_pct_out_of_range_raises(self):
+        df = _make_discount_events()
+        df["discount_pct"] = [1.5, 0.2]
+        with pytest.raises(ValueError, match=r"must be in \[0, 1\]"):
+            DiscountedEventEffect(df_events=df, prefix="promo")
+
+
+class TestDiscountedEventEffectUnits:
+    def test_lever_bounds_native_units(self):
+        effect = DiscountedEventEffect(
+            df_events=_make_discount_events(),
+            prefix="promo",
+            discount_min=0.05,
+            discount_max=0.4,
+        )
+        assert effect.lever_bounds == [(0.05, 0.4), (0.05, 0.4)]
+
+    def test_window_mask_marks_event_dates(self, dates):
+        effect = DiscountedEventEffect(
+            df_events=_make_discount_events(), prefix="promo"
+        )
+        window = effect._window_mask(dates)
+        assert window.dims == ("date", "promo")
+        assert list(window.coords["promo"].values) == ["black_friday", "summer_sale"]
+        in_window = (dates >= pd.Timestamp("2025-03-03")) & (
+            dates <= pd.Timestamp("2025-03-24")
+        )
+        np.testing.assert_array_equal(
+            window.sel(promo="black_friday").values, in_window.astype(float)
+        )
+
+    def test_future_event_warns_zero_revenue(self, dates):
+        df = _make_discount_events()
+        df.loc[1, ["start_date", "end_date"]] = ["2030-01-01", "2030-01-31"]
+        effect = DiscountedEventEffect(df_events=df, prefix="promo")
+
+        class _MMMStub:
+            y = pd.Series(np.ones(len(dates)))
+            X = pd.DataFrame({"date": dates})
[…]
