### 2026-07-30 01f0ee8
chore(deps): bump pymc floor to 6.0.1
The pymc.dims migration uses pmd.Poisson/NegativeBinomial, which landed in pymc 6.0.1, so the oldest-deps CI shard must resolve 6.0.1 instead of 6.0.0.
