+++ pyproject.toml
@@ -33,7 +33,7 @@ dependencies = [
     "pandas",
     "patsy",
     "pydantic>=2.1.0",
-    "pymc>=6.0.0,<6.1.0",
+    "pymc>=6.0.1,<6.1.0",
     "pytensor>=3.0.0,<4.0",
     "scikit-learn>=1.1.1",
     "seaborn>=0.12.2",
