### 2026-07-30 eb36008
fix(bass): satisfiable test prior and transpose without a T coord
- test_save_load_round_trip_scaled_priors: p prior mu=0.02, drops the preliz mass warning (3 of the suite's 4)
- _from_xarray transposes before adding the T coord, so a (product, T) Dataset with no T coord gets the right T length; regression test added
