### 2026-07-30 1f9937a
chore: migrate from arviz.InferenceData to xarray.DataTree
