+++ pymc_marketing/customer_choice/mnl_logit.py
@@ -617,7 +617,7 @@ class MNLogit(ModelBuilder):
             warnings.filterwarnings(
                 "ignore",
                 category=UserWarning,
-                message="The group fit_data is not defined in the InferenceData scheme",
+                message="The group fit_data is not defined in the DataTree scheme",
             )
             self.idata["/fit_data"] = fit_data
 
