### 2026-07-30 53dfbdb
feat: add color_dim to plot_curve and plot_decomposition to FourierBase
- _plot_across_coord, plot_hdi, plot_samples, plot_curve: new
  color_dim parameter for using a dimension as color/legend
  instead of faceting it
- FourierBase.plot_decomposition: thin wrapper around
  plot_curve(color_dim=self.prefix) to show individual sin/cos
  components with HDI + sample lines
- FourierBase.sample_curve: add sum parameter, default True
  (backward compatible), returns per-component curves when False
- Module docstring: two linked .. plot:: blocks — Block 1 plots
  the total curve; Block 2 reuses the 
[…]
