+++ pymc_marketing/mmm/fourier.py
@@ -32,18 +32,33 @@ There are three types of Fourier seasonality transformations available:
     from pymc_marketing.mmm import YearlyFourier
     from pymc_extras.prior import Prior
 
+    seed = sum(map(ord, "Yearly"))
+    rng = np.random.default_rng(seed)
+
     prior = Prior(
         "Normal",
-        mu=[0, 0, -1, 0],
+        mu=[0.5, 0, -1, 0],
         sigma=Prior("Gamma", mu=0.10, sigma=0.1, dims="fourier"),
         dims=("hierarchy", "fourier"),
     )
     yearly = YearlyFourier(n_order=2, prior=prior)
     coords = {"hierarchy": ["A", "B"]}
-    prior = yearly.sample_prior(coords=coords)
+    prior = yearly.sample_prior(coords=coords, random_seed=rng)
     curve = yearly.sample_curve(prior)
     fig, _ = yearly.plot_curve(curve, subplot_kwargs={"ncols": 1})
     fig.suptitle("Yearly Fourier Seasonality")
+    plt.show()
+
+.. plot::
+    :context: close-figs
+
+    # Reuse yearly and prior from the previous example
+    components = yearly.sample_curve(prior, sum=False)
+
+    fig, _ = yearly.plot_decomposition(components)
+    fig.suptitle(
+        "Yearly Fourier seasonality decomposed into basis functions"
+    )
     plt.show()
 
 Examples
@@ -250,7 +265,12 @@ from pytensor.xtensor.type import XTensorVariable
 
 from pymc_marketing.constants import DAYS_IN_MONTH, DAYS_IN_WEEK, DAYS_IN_YEAR
 from pymc_marketing.mmm.dims import XTensorLike
-from pymc_marketing.plot import SelToString, plot_curve, plot_hdi, plot_samples
+from pymc_marketing.plot import (
+    SelToString,
+    plot_curve,
+    plot_hdi,
+    plot_samples,
+)
 from pymc_marketing.serialization import serialization
 
 X_NAME: str = "day"
@@ -519,6 +539,7 @@ class FourierBase(BaseModel):
         parameters: xr.DataTree | xr.Dataset,
         use_dates: bool = False,
         start_date: str | datetime.datetime | None = None,
+        sum: bool = True,
     ) -> xr.DataArray:
         """Create full period of the Fourier seasonality.
 
@@ -532,6 +553,10 @@ class FourierBase(BaseModel):
         start_date : datetime.datetime, optional
             Starting date for the Fourier curve. If not provided and use_dates is True,
             it will be derived from the current year or month. Defaults to None.
+        sum : bool, default True
+            Whether to sum the Fourier components. If ``False``, returns
+            individual components along the ``{prefix}`` dimension
+            (e.g. ``"fourier"``).
 
         Returns
         -------
@@ -568,7 +593,7 @@ class FourierBase(BaseModel):
             name = f"{self.prefix}_trend"
             pmd.Deterministic(
                 name,
-                self.apply(as_xtensor(dayofperiod, dims=(date_dim,))),
+                self.apply(as_xtensor(dayofperiod, dims=(date_dim,)), sum=sum),
             )
 
             return pm.sample_posterior_predictive(
@@ -645,6 +670,91 @@ class FourierBase(BaseModel):
             hdi_kwargs=hdi_kwargs,
             axes=axes,
             same_axes=same_axes,
+            colors=colors,
+            legend=legend,
+            sel_to_string=sel_to_string,
+        )
+
+    def plot_decomposition(
+        self,
+        curve: xr.DataArray,
+        n_samples: int = 10,
+        hdi_probs: float | list[float] | None = None,
[…]
