+++ pymc_marketing/plot.py
@@ -372,6 +372,7 @@ def _plot_across_coord(
     patch: bool = True,
     line: bool = True,
     sel_to_string: SelToString | None = None,
+    color_dim: str | None = None,
 ) -> tuple[plt.Figure, npt.NDArray[Axes]]:
     """Plot data array across coords.
 
@@ -392,10 +393,13 @@ def _plot_across_coord(
 
     data = get_plot_data(curve)
 
-    plot_coords = get_plot_coords(
-        data.coords,
-        non_grid_names=non_grid_names.union({"chain", "draw", "ci_bound"}),
-    )
+    faceting_non_grid = non_grid_names.union({"chain", "draw", "ci_bound"})
+    if color_dim:
+        faceting_non_grid.add(color_dim)
+        color_values = list(data.coords[color_dim].values)
+        color_labels = [str(v) for v in color_values]
+
+    plot_coords = get_plot_coords(data.coords, non_grid_names=faceting_non_grid)
     total_size = get_total_coord_size(plot_coords)
 
     if axes is None and not same_axes:
@@ -403,57 +407,77 @@ def _plot_across_coord(
         subplot_kwargs = {**{"sharey": True, "sharex": True}, **subplot_kwargs}
         set_subplot_kwargs_defaults(subplot_kwargs, total_size)
         fig, axes = plt.subplots(**subplot_kwargs)
-        axes_iter = np.ravel(axes)
+        axes_flat = np.ravel(axes)
         return_axes = axes
 
         create_title = sel_to_string
 
-        create_legend_label = None
     elif axes is not None and same_axes:
         fig = plt.gcf()
-        axes_iter = repeat(axes[0], total_size)  # type: ignore
+        axes_flat = repeat(axes[0], total_size)  # type: ignore
         return_axes = np.array([axes]) if not isinstance(axes, np.ndarray) else axes
 
         def create_title(sel):
             return ""
 
-        create_legend_label = sel_to_string
-
     elif axes is None and same_axes:
         fig, ax = plt.subplots(ncols=1, nrows=1)
-        axes_iter = repeat(ax, total_size)  # type: ignore
+        axes_flat = repeat(ax, total_size)  # type: ignore
         return_axes = np.array([ax])
 
         def create_title(sel):
             return ""
 
-        create_legend_label = sel_to_string
     else:
         fig = plt.gcf()
-        axes_iter = np.ravel(axes)  # type: ignore
+        axes_flat = np.ravel(axes)  # type: ignore
         return_axes = np.array([axes]) if not isinstance(axes, np.ndarray) else axes
 
         create_title = sel_to_string  # type: ignore
 
-        create_legend_label = None
-
-    colors = cast(Iterable[str], colors or generate_colors(n=total_size, start=0))
     plot_kwargs = plot_kwargs or {}
 
-    for color, ax, sel in zip(colors, axes_iter, selections(plot_coords), strict=False):
-        ax = data.pipe(make_selection, sel=sel).pipe(
-            plot_selection,
-            ax=ax,
-            color=color,
-            **plot_kwargs,
+    if color_dim:
+        color_list = cast(
+            Iterable[str], colors or generate_colors(n=len(color_values), start=0)
         )
-        title = create_title(sel)
-        ax.set_title(title)
+        for ax, sel in zip(axes_flat, selections(plot_coords), strict=False):
+            for color, cval in zip(color_list, color_values, strict=False):
+                inner_sel = {**sel, color_dim: cval} if sel else {color_dim: cval}
+                data.pipe(make_selection, sel=inner_sel).pipe(
+                    plot_selection,
+                    ax=ax,
+                    color=color,
+                    **plot_kwargs,
+                )
+            title = create_title(sel)
[…]
