### 2026-07-30 60c6861
docs: add prose linking the two Fourier seasonality plot examples
