### 2026-07-30 e254eb1
docs: clarify that the decomposition is the same curve from the figure above
