### 2026-07-30 ba66e35
fix: update test assertions and remaining references after InferenceData→DataTree migration
