### 2026-07-30 c467b26
feat: add color_dim as legend title in color_dim mode
