### 2026-07-30 61c31ac
Fix MMM-VAR notebook dependency setup
