### 2026-07-30 ecdbb6f
perf: PyTensor rewrite Sum(ElemwiseMul) → Dot for faster MCMC sampling
Introduce local_sum_mul_to_dot, a PyTensor graph rewrite that converts
Sum(ElemwiseMul(ExpandDims(w), X), axis=-1) -> Dot(X, w) during logp
compilation.

The rewrite targets MMM channel, Fourier, and control contribution
computations — replacing the element-wise multiply-then-sum pattern
with a fused BLAS-level dot product.

Key design:
- Registered with use_db_name_as_tag=False — opt-in only
- Activated via RegressionModelBuilder._get_sampling_rewrites()
- MMM returns ['local_sum_mul_to_dot'] from that method
- Applied onl
[…]
