+++ tests/mmm/test_rewrites.py
@@ -1,0 +1,289 @@
+#   Copyright 2022 - 2026 The PyMC Labs Developers
+#
+#   Licensed under the Apache License, Version 2.0 (the "License");
+#   you may not use this file except in compliance with the License.
+#   You may obtain a copy of the License at
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+#   Unless required by applicable law or agreed to in writing, software
+#   distributed under the License is distributed on an "AS IS" BASIS,
+#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#   See the License for the specific language governing permissions and
+#   limitations under the License.
+"""Tests for :mod:`pymc_marketing.mmm.rewrites`.
+
+Uses graph comparison strategies from the PyTensor test suite:
+inspecting the compiled function graph for expected ops.
+"""
+
+import numpy as np
+import pytensor.tensor as pt
+import pytest
+from pytensor.compile import Mode, function
+from pytensor.tensor.math import Dot, Sum
+
+from pymc_marketing.mmm.rewrites import create_sampling_mode
+
+# -- helpers ----------------------------------------------------------------
+
+
+def _has_op(fn, op_cls):
+    """Check if any node in the compiled function is an instance of *op_cls*."""
+    return any(isinstance(node.op, op_cls) for node in fn.maker.fgraph.toposort())
+
+
+def _count_op(fn, op_cls):
+    """Count nodes matching *op_cls* in the compiled function."""
+    return sum(1 for node in fn.maker.fgraph.toposort() if isinstance(node.op, op_cls))
+
+
+# -- graph structure tests --------------------------------------------------
+
+
+class TestGraphTransformation:
+    """Verify the rewrite changes the graph structure: Dot replaces Sum."""
+
+    @pytest.mark.parametrize(
+        "shape",
+        [
+            (10, 3),
+            (52, 8),
+            (156, 20),
+            (10, 5, 3),
+            (52, 5, 8),
+            (156, 10, 15),
+        ],
+        ids=lambda s: f"{'x'.join(str(d) for d in s)}",
+    )
+    def test_dot_appears_sum_disappears(self, shape):
+        """After rewrite: Dot is present, Sum over the target axis is gone."""
+        ndim = len(shape)
+        X = pt.tensor3("X") if ndim == 3 else pt.matrix("X")
+        w = pt.dvector("w")
+        expr = (X * w).sum(axis=-1)
+
+        mode = create_sampling_mode()
+        fn = function([X, w], expr, mode=mode)
+
+        # Dot must appear
+        assert _has_op(fn, Dot), "Rewrite should introduce a Dot op in the graph"
+
+        # Sum must decrease relative to default mode
+        fn_default = function([X, w], expr)
+        assert _count_op(fn, Sum) < _count_op(fn_default, Sum), (
+            "Rewrite should reduce Sum count"
+        )
+
+    def test_default_mode_unchanged(self):
+        """Without the rewrite, the graph preserves Sum/Mul."""
+        X = pt.matrix("X")
+        w = pt.dvector("w")
+        expr = (X * w).sum(axis=-1)
+
+        fn = function([X, w], expr)
+
+        assert _has_op(fn, Sum), "Expected Sum in default-compiled graph"
+        assert not _has_op(fn, Dot), "Did not expect Dot in default-compiled graph"
+
+    def test_output_node_is_dot_with_rewrite(self):
+        """The output-producing node should be a Dot under the rewrite."""
+        X = pt.matrix("X")
+        w = pt.dvector("w")
+        expr = (X * w).sum(axis=-1)
+
+        mode = create_sampling_mode()
[…]
