+++ pymc_marketing/mmm/rewrites.py
@@ -1,0 +1,142 @@
+#   Copyright 2022 - 2026 The PyMC Labs Developers
+#
+#   Licensed under the Apache License, Version 2.0 (the "License");
+#   you may not use this file except in compliance with the License.
+#   You may obtain a copy of the License at
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+#   Unless required by applicable law or agreed to in writing, software
+#   distributed under the License is distributed on an "AS IS" BASIS,
+#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#   See the License for the specific language governing permissions and
+#   limitations under the License.
+
+"""PyTensor rewrites for MMM optimization.
+
+Provides a rewrite that converts
+``Sum(ElemwiseMul(ExpandDims(w), X))`` to ``Dot(X, w)``
+for improved performance during MCMC sampling.
+
+The rewrite is registered with ``use_db_name_as_tag=False``, so it is
+**not** included in the default ``fast_run`` optimization pipeline.
+It must be explicitly enabled — see :func:`create_sampling_mode`.
+"""
+
+import pytensor.tensor as pt
+from pytensor.compile import Mode, get_default_mode, optdb
+from pytensor.graph.rewriting.basic import copy_stack_trace, node_rewriter
+from pytensor.scalar import Mul as ScalarMul
+from pytensor.tensor.elemwise import DimShuffle, Elemwise
+from pytensor.tensor.math import Sum
+
+
+@node_rewriter([Sum])
+def local_sum_mul_to_dot(fgraph, node):
+    """Rewrite ``Sum(ElemwiseMul(ExpandDims(w), X), axis=last) -> Dot(X, w)``.
+
+    Detects the pattern ``(X * expand_dims(w)).sum(axis=-1)`` and replaces
+    it with ``pt.dot(X, w)``. This eliminates the intermediate ``(..., C)``
+    tensor and replaces the element-wise multiply + reduce with a fused
+    BLAS-level operation.
+
+    The rewrite targets the common MMM pattern where channel-level
+    coefficients are broadcast via DimShuffle then summed:
+
+        channel_contribution = (channel_data * beta).sum(dim="channel")
+
+    becomes:
+
+        mu = pt.dot(channel_data, beta)
+
+    Parameters
+    ----------
+    fgraph : FunctionGraph
+        The function graph being rewritten.
+    node : Apply
+        A ``Sum`` node to potentially rewrite.
+
+    Returns
+    -------
+    list[Variable] | None
+        A list with the new output variable, or ``None`` if the rewrite
+        does not apply.
+    """
+    if node.op.axis is None:
+        return None
+
+    [inp] = node.inputs
+    if inp.owner is None:
+        return None
+    if not isinstance(inp.owner.op, Elemwise):
+        return None
+    if not isinstance(inp.owner.op.scalar_op, ScalarMul):
+        return None
+
+    mul_inputs = inp.owner.inputs
+    if len(mul_inputs) != 2:
+        return None
+
+    a, b = mul_inputs
+
+    def _unwrap_weight(var):
+        """Return the inner 1D vector if *var* is a DimShuffle of one."""
+        if var.owner is not None and isinstance(var.owner.op, DimShuffle):
+            inner = var.owner.inputs[0]
+            if inner.type.ndim == 1:
+                return inner
+        return None
+
+    weight_a = _unwrap_weight(a)
+    if weight_a is not None:
+        weight = weight_a
+        data = b
+    else:
+        weight = _unwrap_weight(b)
+        if weight is None:
+            return None
+        data = a
+
+    result = pt.dot(data, weight)
[…]
