+++ pyproject.toml
@@ -224,5 +224,6 @@ dev = [
     "h5py>=3.16.0",
     "mypy>=2.1.0",
     "pandas-stubs>=3.0.3.260530",
+    "pytest>=9.1.1",
     "pyyaml>=6.0.3",
 ]
