### 2026-07-30 7bd1a93
fix: guard rewrite to only apply on single-axis sum over last dim
Add guards preventing the Sum-to-Dot rewrite from firing on non-matching axis patterns:
- Skip multi-axis sums (e.g. axis=(0,1))
- Skip sums where the reduction axis is not the last dim

Also adds two tests for these invariants.
