### 2026-07-31 4cbf321
test: add jax and mlx to linker test matrix
Verified rewrite works on all 4 linkers (cvm, numba, jax, mlx).
JAX and MLX skip gracefully if not installed via pytest.importorskip.
