### 2026-07-31 2caa443
Support multidimensional inputs in to_xarray
