### 2026-07-31 7e1ac23
docs(bass): update notebook to the BassModel workflow (#2770)
* fix(bass): restore Prior objects when loading a saved model

BassModel.load left model_config as plain dicts after the JSON round-trip,
so build_model crashed on any loaded model. Parse the config in __init__
like CLV does, and register Scaled in the TypeRegistry so the scaled
market-potential prior round-trips too.

* fix(bass): keep product dimension when forecasting without observed data

_data_setter filled y_obs with 1-D zeros, so out-of-sample prediction
failed for multi-product models.

* docs(bass): update notebook to the 
[…]
