+++ pymc_marketing/bass/model.py
@@ -155,6 +155,7 @@ from pymc_extras.prior import Censored, Prior, VariableFactory, create_dim_handl
 from pymc_marketing.bass import plotting
 from pymc_marketing.bass.data import to_bass_dataset
 from pymc_marketing.model_builder import ModelBuilder, create_sample_kwargs
+from pymc_marketing.model_config import parse_model_config
 from pymc_marketing.version import __version__
 
 
@@ -490,6 +491,16 @@ class BassModel(ModelBuilder):
     _model_type = "BassModel"
     version = __version__
 
+    def __init__(
+        self,
+        model_config: dict | None = None,
+        sampler_config: dict | None = None,
+    ):
+        super().__init__(model_config=model_config, sampler_config=sampler_config)
+        # Restore Prior objects from the dicts produced by the JSON
+        # round-trip in save/load
+        self.model_config = parse_model_config(self.model_config)
+
     @property
     def default_model_config(self) -> dict:
         """Default model configuration with weakly informative priors."""
@@ -541,8 +552,13 @@ class BassModel(ModelBuilder):
         if "observed" in ds:
             set_data["y_obs"] = ds["observed"].values
         elif "y_obs" in self.model:
-            dtype = self.model["y_obs"].get_value().dtype
-            set_data["y_obs"] = np.zeros(len(new_t), dtype=dtype)
+            old_value = self.model["y_obs"].get_value()
+            dims = self.model.named_vars_to_dims["y_obs"]
+            new_shape = tuple(
+                len(new_t) if d == "T" else size
+                for d, size in zip(dims, old_value.shape, strict=True)
+            )
+            set_data["y_obs"] = np.zeros(new_shape, dtype=old_value.dtype)
         with self.model:
             pm.set_data(set_data, coords={"T": new_t})
 
