### 2026-07-31 302a10a
chore: migrate from arviz.InferenceData to xarray.DataTree (#2782)
* chore: migrate from arviz.InferenceData to xarray.DataTree

* fix: update test assertions and remaining references after InferenceData→DataTree migration
