### 2026-07-31 ef35bcd
Remove merge_group triggers from CI workflows
