+++ .github/workflows/pypi.yml
@@ -4,7 +4,6 @@ on:
     branches: [main, v1.0.0]
   push:
     branches: [main, v1.0.0]
-  merge_group:
   release:
     types: [published]
 
