### 2026-07-31 8bc175a
docs: address JOSS review feedback on documentation and CLV (#2748)
* docs: address JOSS review feedback on documentation

* docs: clarify future_t parameter is always in months

---------

Co-authored-by: Juan Orduz <juanitorduz@gmail.com>
