### 2026-07-31 6cf4191
[pre-commit.ci] pre-commit autoupdate (#2731)
updates:
- [github.com/astral-sh/ruff-pre-commit: v0.15.20 → v0.16.0](https://github.com/astral-sh/ruff-pre-commit/compare/v0.15.20...v0.16.0)
- [github.com/pre-commit/mirrors-mypy: v2.1.0 → v2.3.0](https://github.com/pre-commit/mirrors-mypy/compare/v2.1.0...v2.3.0)

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Juan Orduz <juanitorduz@gmail.com>
