### 2026-07-31 58b23eb
chore(bass): drop the accidental data.py change (merge artifact)
data.py belongs to the notebook PR (now merged into main); an earlier sync merge duplicated the _from_xarray transpose block here. Restore it to main so this PR's diff is only the pymc.dims migration.
