### 2026-08-10 pymc-labs/pymc-marketing#2845 @DanWaxman
[JOSS Review] AI Disclosure in Paper
This is as part of my review for https://github.com/openjournals/joss-reviews/issues/10805.

The paper's AI usage declaration states:
> Generative AI tools were used during paper preparation. OpenCode (v1.0.220) with Claude Opus 4.5 and OpenCode (v1.14.48) with Claude Sonnet 4.6 via GitHub Copilot assisted with gathering information from existing documentation and codebase, drafting text, and incorporating peer reviewer feedback. The PyMC-Marketing software itself was developed by human contributors. All paper content was reviewed, edited, and validated by the human authors.

While I don't think this was purposeful --- and the software has an extensive pre-LLM history + would pass the bar anyways --- I read this statement as saying there was no AI assistance in software development; this conflicts with usage of Claude Code and Cursor Agent, in the git commits at least in January of this year. I think the AI usage declaration would be more transparent if it included that some code has been LLM-assisted, but architectural changes were human-made, etc.
