### 2026-08-10 pymc-labs/pymc-marketing#2846 @DanWaxman
[JOSS Review] Git Size
Again as part of my review on https://github.com/openjournals/joss-reviews/issues/10805. This one is not blocking to me to recommend acceptance, but something I noticed during my review.

The git file for this repository is quite large. This seems largely due to many notebook changes and some committed model weights. This is in no ways insurmountable, but it does make downloading + contributing to the project a bit more annoying.

Feel free to close if you decide against it, but I would perhaps consider a careful `git filter-repo` and maybe moving model weights to `git-lfs` or similar.
