### 2026-08-10 pymc-labs/pymc-marketing#2845 @juanitorduz
Thank you @DanWaxman ! We will address this comment to bring more transparency about the AI usage.
