### 2026-08-10 pymc-labs/pymc-marketing#2845 @juanitorduz
closed by https://github.com/pymc-labs/pymc-marketing/pull/2847
