### 2026-08-12 986f1b7
paper: match bib titles to source websites, protect Gaussian capitalization
Address nkrusch's JOSS review (openjournals/joss-reviews#10805): update the
Bolt and HelloFresh blog-post titles to match the published pages verbatim,
and brace {Gaussian} in the Solin & Särkkä entry so the proof does not
sentence-case it. Rebuild paper.pdf.

Co-Authored-By: Claude Fable 5 <noreply@anthropic.com>
