+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @online{pymclabs2023hellofresh,
   author = {{PyMC Labs}},
-  title = {{Improving the Speed and Accuracy of Bayesian Media Mix Models: A HelloFresh Case Study}},
+  title = {{Improving the Speed and Accuracy of Bayesian Media Mix Models for Marketing Optimization}},
   url = {https://www.pymc-labs.com/blog-posts/reducing-customer-acquisition-costs-how-we-helped-optimizing-hellofreshs-marketing-budget/},
   urldate = {2026-05-14},
   year = {2023}
@@ -8,7 +8,7 @@
 
 @online{pymclabs2023bolt,
   author = {{PyMC Labs}},
-  title = {{Bolt's Success Story with PyMC-Marketing}},
+  title = {{Better budgeting with Bayesian models: Bolt's story with PyMC-Marketing}},
   url = {https://bolt.eu/en/blog/budgeting-with-bayesian-models-pymc-marketing/},
   urldate = {2026-07-20},
   year = {2023}
@@ -98,7 +98,7 @@
 }
 
 @article{solin2020hilbert,
-  title={Hilbert space methods for reduced-rank Gaussian process regression},
+  title={Hilbert space methods for reduced-rank {Gaussian} process regression},
   author={Solin, Arno and S{\"a}rkk{\"a}, Simo},
   journal={Statistics and Computing},
   volume={30},
