### 2026-08-12 pymc-labs/pymc-marketing#2846 @williambdean
Here are some details from a full analysis:

## Current State

### Fresh Clone Size

| Component | Size |
|-----------|------|
| `.git` directory | 6.3 GB |
| Working tree (tracked files) | ~415 MB |
| **Total clone** | **~6.7 GB** |

### What's Bloating Git

The top 50 largest objects in git history are entirely `.ipynb` and `.nc` files.

**Notebooks (72 files, 264 MB working tree):**

| Notebook | Commits | Size | Cumulative history |
|----------|---------|------|--------------------|
| `mmm_example.ipynb` | 9 | 12 MB | High |
| `mmm_migration_guide.ipynb` | 9 | 1 MB | High |
| `mmm_gam_options.ipynb` | 8 | 6.5 MB | High |
| `mmm_budget_allocation_example.ipynb` | 8 | 3.8 MB | High |
| `mmm_case_study.ipynb` | 7 | 20 MB | **~700 MB** (38 blob versions) |

- 94% of notebooks have been modified in multiple commits (up to 9 each)
- Every notebook edit stores a full 10-20 MB binary copy in `.git` forever

**NetCDF files (4 tracked, 91 MB working tree):**

- `multidimensional_model.nc` — 91 MB, still in the working tree
- `model.nc` — 7 historical versions at **51-77 MB each** (~500 MB total), deleted from tree but lives in history
- `budget_optimizer_model.nc` — one 65 MB version, also in history

### What's NOT Worth Addressing

| File Type | Count | Size | Why excluded |
|-----------|-------|------|-------------|
| `.csv` data | 21 | 1.7 MB | All committed once, never modified. Too small for LFS overhead |
| `.png` / `.jpg` images | 43 | ~11 MB | Static doc assets, rarely change. 0.2% of the problem |
| `.po` translations | 1,804 | 12 MB | Text files, tiny individually, git handles them efficiently |

## Options Considered

### Option A: LFS for `.ipynb` + `.nc` **(Recommended)**

**What it does:** Track both file types in LFS, rewrite history to purge old binary blobs.

| `.git` after | Clone after | Complexity |
|-------------|-------------|------------|
| ~500 MB | ~915 MB | Medium (5 config files + RTD workaround) |

- **Pros:** Solves both problems permanently. CI `test.yml` gets 87% faster checkout. No notebook output changes. No regrowth.
- **Cons:** Needs `.readthedocs.yaml` with `post_checkout` git-lfs workaround. Team coordination for history rewrite and contributor setup.

### Option B: LFS for `.nc` only

| `.git` after | Clone after | Complexity |
|-------------|-------------|------------|
| ~5.8 GB | ~6.2 GB | Low |

- **Pros:** Simple, no RTD workaround needed, no notebook impact.
- **Cons:** Biggest problem (notebook churn re-bloating `.git` on every edit) remains completely unsolved.

### Option C: LFS for `.ipynb` only

| `.git` after | Clone after | Complexity |
|-------------|-------------|------------|
| ~550 MB | ~965 MB | Medium (RTD workaround still needed) |

- **Pros:** Solves the notebook bloat (the main ongoing problem).
- **Cons:** 91 MB `multidimensional_model.nc` stays. Loses the simplicity of a single rule in `.gitattributes`.

### Option D: LFS for `.nc` + only large `.ipynb` (> 5 MB)

| `.git` after | Clone after | Complexity |
|-------------|-------------|------------|
| ~550 MB | ~960 MB | Medium |

- **Pros:** Smaller blast radius, fewer notebooks in LFS.
- **Cons:** Arbitrary threshold. Small notebooks regrow if they accumulate outputs and cross the line. No clear maintenance story.

### Option E: History rewrite only (no LFS)

| `.git` after | Clone after | Complexity |
|-------------|-------------|------------|
| ~500 MB (but regrows) | ~6.7 GB (regrows) | High (one-time, no ongoing benefit) |

- **Pros:** No LFS dependency, no workflow changes.
- **Cons:** Every subsequent notebook edit stores a full binary copy again. `.git` creeps back to GBs over time. Kicks the can down the road.

### Ruled Out

| Approach | Why ruled out |
|----------|--------------|
| Strip notebook outputs + re-execute in CI | MMM notebooks are already excluded from `test_notebook.yml` because they are too slow. RTD/docs would go from minutes to hours. Outputs are critical for docs users. |
| Remove `.nc` files and regenerate from notebooks | Same problem — re-executing MMM notebooks is prohibitively slow. |
| LFS for CSV files | All 21 CSVs (1.7 MB) were committed once, never changed. Not worth the overhead. |
| LFS for PNG/JPG images | 43 files at 11 MB total, 0.2% of the bloat. Not worth adding another LFS filter rule. |

## Recommended: Option A

Track `.ipynb` and `.nc` in LFS, rewrite history, one-time team coordination.

```gitattributes
*.ipynb filter=lfs diff=lfs merge=lfs -text
*.nc filter=lfs diff=lfs merge=lfs -text
```

```bash
git lfs migrate import --everything --include="*.ipynb,*.nc"
```

**Notebook outputs are preserved.** LFS stores the full notebook including embedded images and outputs. `git lfs pull` materializes them on clone. The docs build (`myst-nb` with `nb_execution_excludepatterns = ["*.ipynb"]`) reads pre-computed notebooks as-is, same as today.

## ReadTheDocs Impact

RTD has no native LFS support, but a well-established workaround uses `build.jobs.post_checkout` to install git-lfs and pull files before Sphinx builds:
- [readthedocs/readthedocs.org#8288](https://github.com/readthedocs/readthedocs.org/issues/8288) (SecureDrop)
- [readthedocs/readthedocs.org#12479](https://github.com/readthedocs/readthedocs.org/issues/12479) (Tethys)

```yaml
# .readthedocs.yaml (does not exist yet, needs to be created)
version: 2
build:
  os: ubuntu-24.04
  tools:
    python: "3.12"
  jobs:
    post_checkout:
      - wget https://github.com/git-lfs/git-lfs/releases/download/v3.7.0/git-lfs-linux-amd64-v3.7.0.tar.gz
      - tar xvfz git-lfs-linux-amd64-v3.7.0.tar.gz
      - ./git-lfs-3.7.0/git-lfs install
      - ./git-lfs-3.7.0/git-lfs fetch
      - ./git-lfs-3.7.0/git-lfs checkout
sphinx:
  configuration: docs/source/conf.py
  fail_on_warning: true
python:
  install:
    - method: uv
      command: sync
      groups:
        - docs
```

## CI / Dev Workflow Changes

### Files to change (6)

| File | Change | Why |
|------|--------|-----|
| `.gitattributes` | Add LFS rules | Core LFS tracking |
| `.readthedocs.yaml` | Create (see above) | RTD has no native LFS |
| `.github/workflows/docs.yml` | Add `lfs: true` to checkout | Explicit LFS pull (already default for v7, but be safe) |
| `.github/workflows/test_notebook.yml` | Add `lfs: true` to checkout | Notebook runner needs real content, not LFS pointers |
| `.github/workflows/test.yml` | Add `lfs: false` to checkout | Doesn't touch notebooks/NC — skip LFS for 87% faster checkout |
| `Makefile` | Add `git lfs pull` to `html`, `run_notebooks` targets | Local dev needs content before building docs |

### No changes needed

| What | Why |
|------|-----|
| Test code (`tests/mmm/test_mmm.py`, etc.) | All `.nc` writes use `tmp_path` — LFS is transparent at filesystem level |
| `scripts/run_notebooks/runner.py` | Reads notebooks from disk — after `git lfs pull`, files are real |
| All other workflows (pypi, uml, link-checker, etc.) | None interact with `.ipynb` or `.nc` |

### Dev lifecycle

No change to daily notebook editing. LFS operates transparently at `git add` time:

- **Editing/saving notebooks in Jupyter:** Identical to today
- **`git add` / `git commit` / `git push`:** Same commands, LFS clean filter handles storage transparently
- **Fresh clone:** `git lfs install` once globally (one-time), then `git clone` auto-pulls LFS files
- **After `git pull`:** Smudge filter auto-materializes updated files

## CI Transfer Speed Improvement

| Workflow | Jobs/run | Needs LFS? | Before (per job) | After (per job) | Reduction |
|----------|----------|-----------|-------------------|------------------|-----------|
| `test.yml` | 16 | **No** | ~500-800 MB | **~50-100 MB** | **~87%** |
| `docs.yml` | 1 | **Yes** | ~500-800 MB | ~405-455 MB | ~35% |
| `test_notebook.yml` | 14 | **Yes** | ~500-800 MB | ~405-455 MB | ~35% |

`test.yml` gets the biggest win by skipping LFS entirely — it never touches notebooks or NC files.

## Expected Outcome

| | Before | After |
|---|--------|-------|
| `.git` directory | 6.3 GB | ~500 MB |
| Clone total | ~6.7 GB | **~915 MB** |
| `test.yml` CI checkout | ~500-800 MB/job | ~50-100 MB/job |

## Execution Steps

1. Add LFS rules to `.gitattributes`
2. Create `.readthedocs.yaml` with `post_checkout` git-lfs workaround + uv build config
3. Add explicit `lfs: true` to `docs.yml` and `test_notebook.yml`; `lfs: false` to `test.yml`
4. Add `git lfs pull` to `Makefile` `html` and `run_notebooks` targets
5. Run `git lfs migrate import --everything --include="*.ipynb,*.nc"`
6. Force push all branches/tags
7. Team coordination (see below)

## Team Coordination

### New Contributors (after migration)

After migration, new contributors need a one-time git-lfs setup before cloning. This should be documented in `CONTRIBUTING.md` and any onboarding/setup guides:

| Platform | Command |
|----------|---------|
| **macOS** | `brew install git-lfs && git lfs install` |
| **Ubuntu/Debian** | `sudo apt install git-lfs && git lfs install` |
| **Windows** | `git lfs install` (bundled with Git for Windows) |

After that, `git clone` works as normal — LFS auto-pulls notebook and NC file content via the smudge filter.

### Existing Contributors (at migration time)

Since `git lfs migrate` rewrites every commit SHA, existing clones become orphaned. All existing contributors need to:

**Step 1 — Install git-lfs (one-time):**

| Platform | Command |
|----------|---------|
| **macOS** | `brew install git-lfs && git lfs install` |
| **Ubuntu/Debian** | `sudo apt install git-lfs && git lfs install` |
| **Windows** | `git lfs install` (bundled with Git for Windows) |

**Step 2 — Recover existing clone:**

| Situation | Action |
|-----------|--------|
| **Clean clone (no unmerged work)** | `git fetch --all && git reset --hard origin/main` |
| **Clone with uncommitted changes** | `git stash && git fetch --all && git reset --hard origin/main && git stash pop` |
| **Git worktrees** | Delete and re-create. First reset the main clone, then: `git worktree remove <path> && git worktree add <path> <branch>` |
| **Open pull requests** | Rebase onto the rewritten base: `git fetch origin && git rebase origin/main` (may need conflict resolution) |

### What "Rebasing All PRs" Actually Means

The impact depends on what each PR touches. After `git lfs migrate` rewrites history, `.ipynb` and `.nc` files in old commits become LFS pointers, but `.py`, `.csv`, and all other files are untouched:

| Type | Rebase behavior |
|------|----------------|
| **Python/config only** | Clean — `git fetch && git rebase origin/main` just works |
| **Adds new `.ipynb`** | Clean — LFS filter auto-converts the new file during cherry-pick |
| **Modifies existing `.ipynb`** | **Manual work needed** — old binary patches don't apply to new LFS pointers. Author needs to re-apply their notebook changes after rebasing. Python changes in the same PR rebase cleanly. |

The count of affected PRs depends on what's open at migration time. The ideal window is right after a merge batch, when few-to-zero notebook-modifying PRs are in flight.

### Recommended Coordination Window

1. Announce the migration date in the team channel at least 48 hours ahead
2. Ask contributors to merge or close any low-priority PRs before the window
3. Execute the migration during a low-activity period (e.g., evening or weekend)
4. After force push, post the step-by-step recovery guide in the team channel
5. Update `CONTRIBUTING.md` and onboarding docs with the git-lfs prerequisite
