### 2026-08-12 pymc-labs/pymc-marketing#2846 @DanWaxman
A quick word of caution, from (painful) experience... GitHub's LFS stores objects forever, which is pretty annoying. So I still wouldn't use GitHub's LFS (but maybe other hosted LFSs are fine) for files that you expect to change often.
