### 2026-08-12 pymc-labs/pymc-marketing#2846 @williambdean
Thanks for words of advice, Dan ! 🙏
