### 2026-08-20 #10805 @nkrusch
Hi @juanitorduz and @williambdean

just a gentle reminder, please share the archive DOI and release tag, so I can move this submission forward.
