### 2026-08-24 43db76f
Bump version to 1.1.0
Prepares the 1.1.0 release, which will provide the release tag and
archive DOI requested by the JOSS editor in openjournals/joss-reviews#10805.

Minor bump: several feat commits have landed on main since 1.0.0.
