### 2026-08-27 #10805 @juanitorduz
@nkrusch, apologies for the delay. We just cut a release:

https://github.com/pymc-labs/pymc-marketing/releases/tag/1.1.0

and updated the Zenodo entry https://zenodo.org/records/22125692

Please let us know if you need anything else from our side.

Thanks :)

FYI @williambdean
