### 2026-08-27 #10805 @nkrusch
@juanitorduz I need a the following fixes in the Zenodo meta data:

* The version is `https://github.com/pymc-labs/pymc-marketing/releases/tag/1.1.0` --- it should be `1.1.0`
* The version is Creative Commons Attribution 4.0 International --- change the license to to Apache-2.0
