### 2026-08-27 #10805 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
