### 2026-08-27 #10805 @nkrusch
@juanitorduz @williambdean,

I am now done with my tasks. From here the submission moves to the JOSS track editor who will do a final round of checks.
Please wait for updates from the track editor.
